@@ -4,32 +4,35 @@ Defines some useful types for this library. Should probably be cleaned up before
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Protocol, TypeVar, Union
+from typing import TYPE_CHECKING, Literal, Protocol, TypeVar
 
-from .compat import (
-    H5Array,
-    H5Group,
-    ZarrArray,
-    ZarrGroup,
-)
+from .compat import H5Array, H5Group, ZarrArray, ZarrGroup
 from .typing import RWAble
 
 if TYPE_CHECKING:
     from collections.abc import Mapping
     from typing import Any, TypeAlias
 
-    from ._io.specs.registry import DaskReader, IOSpec, Reader, Writer
-    from .compat import DaskArray
+    from ._io.specs.registry import (
+        IOSpec,
+        LazyDataStructures,
+        LazyReader,
+        Reader,
+        Writer,
+    )
 
 __all__ = [
     "ArrayStorageType",
     "GroupStorageType",
     "StorageType",
+    "_ReadInternal",
+    "_ReadLazyInternal",
+    "_WriteInternal",
 ]
 
-ArrayStorageType: TypeAlias = Union[ZarrArray, H5Array]
-GroupStorageType: TypeAlias = Union[ZarrGroup, H5Group]
-StorageType: TypeAlias = Union[ArrayStorageType, GroupStorageType]
+ArrayStorageType: TypeAlias = ZarrArray | H5Array
+GroupStorageType: TypeAlias = ZarrGroup | H5Group
+StorageType: TypeAlias = ArrayStorageType | GroupStorageType
 
 # NOTE: If you change these, be sure to update `autodoc_type_aliases` in docs/conf.py!
 ContravariantRWAble = TypeVar("ContravariantRWAble", bound=RWAble, contravariant=True)
@@ -44,10 +47,10 @@ class _ReadInternal(Protocol[SCon, CovariantRWAble]):
     def __call__(self, elem: SCon, *, _reader: Reader) -> CovariantRWAble: ...
 
 
-class _ReadDaskInternal(Protocol[SCon]):
+class _ReadLazyInternal(Protocol[SCon]):
     def __call__(
-        self, elem: SCon, *, _reader: DaskReader, chunks: tuple[int, ...] | None = None
-    ) -> DaskArray: ...
+        self, elem: SCon, *, _reader: LazyReader, chunks: tuple[int, ...] | None = None
+    ) -> LazyDataStructures: ...
 
 
 class Read(Protocol[SCon, CovariantRWAble]):
@@ -60,16 +63,16 @@ class Read(Protocol[SCon, CovariantRWAble]):
             The element to read from.
         Returns
         -------
-            The element read from the store.
+        The element read from the store.
         """
         ...
 
 
-class ReadDask(Protocol[SCon]):
+class ReadLazy(Protocol[SCon]):
     def __call__(
         self, elem: SCon, *, chunks: tuple[int, ...] | None = None
-    ) -> DaskArray:
-        """Low-level reading function for a dask element.
+    ) -> LazyDataStructures:
+        """Low-level reading function for a lazy element.
 
         Parameters
         ----------
@@ -79,7 +82,7 @@ class ReadDask(Protocol[SCon]):
             The chunk size to be used.
         Returns
         -------
-            The dask element read from the store.
+        The lazy element read from the store.
         """
         ...
 
@@ -137,7 +140,7 @@ class ReadCallback(Protocol[SCo, InvariantRWAble]):
         Params
         ------
         read_func
-            :func:`anndata.read_elem` function to call to read the current element given the ``iospec``.
+            :func:`anndata.io.read_elem` function to call to read the current element given the ``iospec``.
         elem_name
             The key to read in from the group.
         elem
@@ -147,7 +150,7 @@ class ReadCallback(Protocol[SCo, InvariantRWAble]):
 
         Returns
         -------
-            The element read from the store.
+        The element read from the store.
         """
         ...
 
@@ -170,7 +173,7 @@ class WriteCallback(Protocol[InvariantRWAble]):
         Params
         ------
         write_func
-            :func:`anndata.write_elem` function to call to read the current element given the ``iospec``.
+            :func:`anndata.io.write_elem` function to call to read the current element given the ``iospec``.
         store
             The store to which `elem` should be written.
         elem_name
@@ -183,3 +186,19 @@ class WriteCallback(Protocol[InvariantRWAble]):
             Keyword arguments to be passed to a library-level io function, like `chunks` for :doc:`zarr:index`.
         """
         ...
+
+
+AnnDataElem = Literal[
+    "obs",
+    "var",
+    "obsm",
+    "varm",
+    "obsp",
+    "varp",
+    "layers",
+    "X",
+    "raw",
+    "uns",
+]
+
+Join_T = Literal["inner", "outer"]
