@@ -1,18 +1,48 @@
-#!python3
+#!/usr/bin/env python3
+# /// script
+# dependencies = [ "towncrier", "packaging" ]
+# ///
 from __future__ import annotations
 
 import argparse
+import re
 import subprocess
+from functools import cache
 from typing import TYPE_CHECKING
 
-from packaging import version
+from packaging.version import Version
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
 
 
+class BumpVersion(Version):
+    def __init__(self, version: str) -> None:
+        super().__init__(version)
+
+        if len(self.release) != 3:
+            msg = f"{version} must contain major, minor, and patch version."
+            raise argparse.ArgumentTypeError(msg)
+
+        base_branch = get_base_branch()
+        patch_branch_pattern = re.compile(r"\d+\.\d+\.x")
+        if self.micro != 0 and not patch_branch_pattern.fullmatch(base_branch):
+            msg = (
+                f"{version} is a patch release, but "
+                f"you are trying to release from a non-patch release branch: {base_branch}."
+            )
+            raise argparse.ArgumentTypeError(msg)
+
+        if self.micro == 0 and base_branch != "main":
+            msg = (
+                f"{version} is a minor or major release, "
+                f"but you are trying to release not from main: {base_branch}."
+            )
+            raise argparse.ArgumentTypeError(msg)
+
+
 class Args(argparse.Namespace):
-    version: str
+    version: BumpVersion
     dry_run: bool
 
 
@@ -28,7 +58,7 @@ def parse_args(argv: Sequence[str] | None = None) -> Args:
     )
     parser.add_argument(
         "version",
-        type=str,
+        type=BumpVersion,
         help=(
             "The new version for the release must have at least three parts, like `major.minor.patch` and no `major.minor`. "
             "It can have a suffix like `major.minor.patch.dev0` or `major.minor.0rc1`."
@@ -40,11 +70,6 @@ def parse_args(argv: Sequence[str] | None = None) -> Args:
         action="store_true",
     )
     args = parser.parse_args(argv, Args())
-    if len(version.Version(args.version).release) != 3:
-        raise ValueError(
-            f"Version argument {args.version} must contain major, minor, and patch version."
-        )
-    version.parse(args.version)  # validate
     return args
 
 
@@ -57,15 +82,8 @@ def main(argv: Sequence[str] | None = None) -> None:
     )
 
     # Check if we are on the main branch to know if we need to backport
-    base_branch = subprocess.run(
-        ["git", "rev-parse", "--abbrev-ref", "HEAD"],
-        capture_output=True,
-        text=True,
-        check=True,
-    ).stdout.strip()
-    pr_description = (
-        "" if base_branch == "main" else "@meeseeksmachine backport to main"
-    )
+    base_branch = get_base_branch()
+    pr_description = "" if base_branch == "main" else "@meeseeksdev backport to main"
     branch_name = f"release_notes_{args.version}"
 
     # Create a new branch + commit
@@ -77,14 +95,7 @@ def main(argv: Sequence[str] | None = None) -> None:
     # push
     if not args.dry_run:
         subprocess.run(
-            [
-                "git",
-                "push",
-                "--set-upstream",
-                "origin",
-                branch_name,
-            ],
-            check=True,
+            ["git", "push", "--set-upstream", "origin", branch_name], check=True
         )
     else:
         print("Dry run, not pushing")
@@ -95,15 +106,11 @@ def main(argv: Sequence[str] | None = None) -> None:
             "gh",
             "pr",
             "create",
-            "--base",
-            base_branch,
-            "--title",
-            pr_title,
-            "--body",
-            pr_description,
-            "--label",
-            "skip-gpu-ci",
-            *(["--label", "no milestone"] if base_branch == "main" else []),
+            f"--base={base_branch}",
+            f"--title={pr_title}",
+            f"--body={pr_description}",
+            "--label=skip-gpu-ci",
+            *(["--label=no milestone"] if base_branch == "main" else []),
             *(["--dry-run"] if args.dry_run else []),
         ],
         check=True,
@@ -116,6 +123,16 @@ def main(argv: Sequence[str] | None = None) -> None:
         )
     else:
         print("Dry run, not merging")
+
+
+@cache
+def get_base_branch():
+    return subprocess.run(
+        ["git", "rev-parse", "--abbrev-ref", "HEAD"],
+        capture_output=True,
+        text=True,
+        check=True,
+    ).stdout.strip()
 
 
 if __name__ == "__main__":
