@@ -22,13 +22,13 @@ sys.path[:0] = [str(_extension_dir)]
 # General information
 project = "anndata"
 author = f"{project} developers"
-copyright = f"{datetime.now():%Y}, {author}"
+copyright = f"{datetime.now():%Y}, scverse"
 release = version = metadata.version("anndata")
 
 # default settings
 templates_path = ["_templates"]
 html_static_path = ["_static"]
-source_suffix = {".rst": "restructuredtext", ".md": "markdown"}
+source_suffix = {".rst": "restructuredtext", ".md": "myst-nb"}
 master_doc = "index"
 default_role = "literal"
 exclude_patterns = [
@@ -44,7 +44,7 @@ exclude_patterns = [
 pygments_style = "sphinx"
 
 extensions = [
-    "myst_parser",
+    "myst_nb",
     "sphinx_copybutton",
     "sphinx.ext.autodoc",
     "sphinx.ext.intersphinx",
@@ -60,27 +60,25 @@ extensions = [
     "sphinxext.opengraph",
     "scanpydoc",  # needs to be before linkcode
     "sphinx.ext.linkcode",
-    "nbsphinx",
     "IPython.sphinxext.ipython_console_highlighting",
-    "patch_sphinx_toolbox_autoprotocol",  # internal extension
     "sphinx_toolbox.more_autodoc.autoprotocol",
-    *(  # other internal extensions
-        p.stem
-        for p in _extension_dir.glob("*.py")
-        if p.stem != "patch_sphinx_toolbox_autoprotocol"
-    ),
+    *(p.stem for p in _extension_dir.glob("*.py")),
 ]
 myst_enable_extensions = [
     "html_image",  # So README.md can be used on github and sphinx docs
+    "colon_fence",
+    "dollarmath",
 ]
 myst_heading_anchors = 3
+nb_execution_mode = "off"
 
 # Generate the API documentation when building
 autosummary_generate = True
 autodoc_member_order = "bysource"
+autodoc_mock_imports = ["torch"]
+# autodoc_default_flags = ['members']
 issues_github_path = "scverse/anndata"
 rtd_links_prefix = PurePosixPath("src")
-# autodoc_default_flags = ['members']
 napoleon_google_docstring = False
 napoleon_numpy_docstring = True
 napoleon_include_init_with_doc = False
@@ -90,31 +88,11 @@ napoleon_custom_sections = [("Params", "Parameters")]
 typehints_defaults = "braces"
 todo_include_todos = False
 nitpicky = True  # Report broken links
-nitpick_ignore = [
-    ("py:class", "scipy.sparse.base.spmatrix"),
-    ("py:meth", "pandas.DataFrame.iloc"),
-    ("py:meth", "pandas.DataFrame.loc"),
-    ("py:class", "anndata._core.views.ArrayView"),
+nitpick_ignore = [  # APIs without an intersphinx entry
+    # This API isn’t actually documented
     ("py:class", "anndata._core.raw.Raw"),
-    *[
-        ("py:class", f"anndata._core.aligned_mapping.{cls}{kind}")
-        for cls in "Layers AxisArrays PairwiseArrays".split()
-        for kind in ["", "View"]
-    ],
-    # TODO: sphinx’ builtin autodoc.typehints extension isn’t handled by `qualname_overrides` yet
-    # https://github.com/theislab/scanpydoc/issues/140
-    ("py:class", "h5py._hl.group.Group"),
-    ("py:class", "h5py._hl.dataset.Dataset"),
-    # for experimental callback exports
+    # TODO: remove zappy support; the zappy repo is archived
     ("py:class", "anndata.compat.ZappyArray"),
-    ("py:class", "anndata.compat.DaskArray"),
-    ("py:class", "anndata.compat.CupyArray"),
-    ("py:class", "anndata.compat.CupySparseMatrix"),
-    ("py:class", "numpy.ma.core.MaskedArray"),
-    ("py:class", "dask.array.core.Array"),
-    ("py:class", "awkward.highlevel.Array"),
-    ("py:class", "anndata._core.sparse_dataset.BaseCompressedSparseDataset"),
-    ("py:obj", "numpy._typing._array_like._ScalarType_co"),
 ]
 
 
@@ -124,27 +102,49 @@ def setup(app: Sphinx):
 
 
 intersphinx_mapping = dict(
-    h5py=("https://docs.h5py.org/en/latest/", None),
-    hdf5plugin=("https://hdf5plugin.readthedocs.io/en/latest/", None),
-    loompy=("https://linnarssonlab.org/loompy/", None),
-    numpy=("https://numpy.org/doc/stable/", None),
-    pandas=("https://pandas.pydata.org/pandas-docs/stable/", None),
+    awkward=("https://awkward-array.org/doc/stable", None),
+    cupy=("https://docs.cupy.dev/en/stable", None),
+    dask=("https://docs.dask.org/en/stable", None),
+    h5py=("https://docs.h5py.org/en/latest", None),
+    hdf5plugin=("https://hdf5plugin.readthedocs.io/en/latest", None),
+    loompy=("https://linnarssonlab.org/loompy", None),
+    numpy=("https://numpy.org/doc/stable", None),
+    pandas=("https://pandas.pydata.org/pandas-docs/stable", None),
     python=("https://docs.python.org/3", None),
-    scipy=("https://docs.scipy.org/doc/scipy/", None),
-    sklearn=("https://scikit-learn.org/stable/", None),
+    scipy=("https://docs.scipy.org/doc/scipy", None),
+    sklearn=("https://scikit-learn.org/stable", None),
     zarr=("https://zarr.readthedocs.io/en/stable/", None),
-    xarray=("https://xarray.pydata.org/en/stable/", None),
-    dask=("https://docs.dask.org/en/stable/", None),
+    xarray=("https://docs.xarray.dev/en/stable", None),
 )
 qualname_overrides = {
     "h5py._hl.group.Group": "h5py.Group",
     "h5py._hl.files.File": "h5py.File",
     "h5py._hl.dataset.Dataset": "h5py.Dataset",
     "anndata._core.anndata.AnnData": "anndata.AnnData",
+    **{
+        f"anndata._core.aligned_mapping.{cls}{kind}": "collections.abc.Mapping"
+        for cls in "Layers AxisArrays PairwiseArrays".split()
+        for kind in ["", "View"]
+    },
     "anndata._types.ReadCallback": "anndata.experimental.ReadCallback",
     "anndata._types.WriteCallback": "anndata.experimental.WriteCallback",
     "anndata._types.Read": "anndata.experimental.Read",
     "anndata._types.Write": "anndata.experimental.Write",
+    "zarr.core.array.Array": "zarr.Array",
+    "zarr.core.group.Group": "zarr.Group",
+    # Buffer is not yet exported, so the buffer class registry is the closest thing
+    "zarr.core.buffer.core.Buffer": "zarr.registry.Registry",
+    "zarr.storage._common.StorePath": "zarr.storage.StorePath",
+    "anndata.compat.DaskArray": "dask.array.Array",
+    "anndata.compat.CupyArray": "cupy.ndarray",
+    "anndata.compat.CupySparseMatrix": "cupyx.scipy.sparse.spmatrix",
+    "awkward.highlevel.Array": "ak.Array",
+    "numpy.int64": ("py:attr", "numpy.int64"),
+    "pandas.DataFrame.iloc": ("py:attr", "pandas.DataFrame.iloc"),
+    "pandas.DataFrame.loc": ("py:attr", "pandas.DataFrame.loc"),
+    # should be fixed soon: https://github.com/tox-dev/sphinx-autodoc-typehints/pull/516
+    "types.EllipsisType": ("py:data", "types.EllipsisType"),
+    "pathlib._local.Path": "pathlib.Path",
 }
 autodoc_type_aliases = dict(
     NDArray=":data:`~numpy.typing.NDArray`",
