@@ -8,10 +8,10 @@ import warnings
 from collections import OrderedDict
 from collections.abc import Mapping, MutableMapping, Sequence
 from copy import copy, deepcopy
-from functools import partial
+from functools import partial, singledispatch
 from pathlib import Path
 from textwrap import dedent
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, cast
 
 import h5py
 import numpy as np
@@ -22,9 +22,11 @@ from pandas.api.types import infer_dtype
 from scipy import sparse
 from scipy.sparse import issparse
 
+from anndata._warnings import ImplicitModificationWarning
+
 from .. import utils
 from .._settings import settings
-from ..compat import DaskArray, SpArray, ZarrArray, _move_adj_mtx
+from ..compat import CSArray, DaskArray, ZarrArray, _move_adj_mtx, old_positionals
 from ..logging import anndata_logger as logger
 from ..utils import (
     axis_len,
@@ -41,7 +43,6 @@ from .raw import Raw
 from .sparse_dataset import BaseCompressedSparseDataset, sparse_dataset
 from .storage import coerce_array
 from .views import (
-    DataFrameView,
     DictView,
     _resolve_idxs,
     as_view,
@@ -50,39 +51,14 @@ from .views import (
 if TYPE_CHECKING:
     from collections.abc import Iterable
     from os import PathLike
-    from typing import Any, Literal
+    from typing import Any, ClassVar, Literal
 
-    from ..typing import ArrayDataStructureType
+    from zarr.storage import StoreLike
+
+    from ..compat import Index1D
+    from ..typing import XDataType
     from .aligned_mapping import AxisArraysView, LayersView, PairwiseArraysView
-    from .index import Index, Index1D
-
-
-# for backwards compat
-def _find_corresponding_multicol_key(key, keys_multicol):
-    """Find the corresponding multicolumn key."""
-    for mk in keys_multicol:
-        if key.startswith(mk) and "of" in key:
-            return mk
-    return None
-
-
-# for backwards compat
-def _gen_keys_from_multicol_key(key_multicol, n_keys):
-    """Generates single-column keys from multicolumn key."""
-    keys = [f"{key_multicol}{i + 1:03}of{n_keys:03}" for i in range(n_keys)]
-    return keys
-
-
-def _check_2d_shape(X):
-    """\
-    Check shape of array or sparse matrix.
-
-    Assure that X is always 2D: Unlike numpy we always deal with 2D arrays.
-    """
-    if X.dtype.names is None and len(X.shape) != 2:
-        raise ValueError(
-            f"X needs to be 2-dimensional, not {len(X.shape)}-dimensional."
-        )
+    from .index import Index
 
 
 class AnnData(metaclass=utils.DeprecationMixinMeta):
@@ -134,15 +110,15 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
     See Also
     --------
-    read_h5ad
-    read_csv
-    read_excel
-    read_hdf
-    read_loom
-    read_zarr
-    read_mtx
-    read_text
-    read_umi_tools
+    io.read_h5ad
+    io.read_csv
+    io.read_excel
+    io.read_hdf
+    io.read_loom
+    io.read_zarr
+    io.read_mtx
+    io.read_text
+    io.read_umi_tools
 
     Notes
     -----
@@ -218,22 +194,35 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         var={"var_names", "col_names", "index"},
     )
 
+    _accessors: ClassVar[set[str]] = set()
+
+    @old_positionals(
+        "obsm",
+        "varm",
+        "layers",
+        "raw",
+        "dtype",
+        "shape",
+        "filename",
+        "filemode",
+        "asview",
+    )
     def __init__(
         self,
-        X: np.ndarray | sparse.spmatrix | pd.DataFrame | None = None,
+        X: XDataType | pd.DataFrame | None = None,
         obs: pd.DataFrame | Mapping[str, Iterable[Any]] | None = None,
         var: pd.DataFrame | Mapping[str, Iterable[Any]] | None = None,
         uns: Mapping[str, Any] | None = None,
+        *,
         obsm: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
         varm: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
-        layers: Mapping[str, np.ndarray | sparse.spmatrix] | None = None,
+        layers: Mapping[str, XDataType] | None = None,
         raw: Mapping[str, Any] | None = None,
         dtype: np.dtype | type | str | None = None,
         shape: tuple[int, int] | None = None,
-        filename: PathLike | None = None,
+        filename: PathLike[str] | str | None = None,
         filemode: Literal["r", "r+"] | None = None,
         asview: bool = False,
-        *,
         obsp: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
         varp: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
         oidx: Index1D | None = None,
@@ -245,7 +234,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 raise_value_error_if_multiindex_columns(attr, key)
         if asview:
             if not isinstance(X, AnnData):
-                raise ValueError("`X` has to be an AnnData object.")
+                msg = "`X` has to be an AnnData object."
+                raise ValueError(msg)
             self._init_as_view(X, oidx, vidx)
         else:
             self._init_as_actual(
@@ -267,19 +257,22 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
     def _init_as_view(self, adata_ref: AnnData, oidx: Index, vidx: Index):
         if adata_ref.isbacked and adata_ref.is_view:
-            raise ValueError(
+            msg = (
                 "Currently, you cannot index repeatedly into a backed AnnData, "
                 "that is, you cannot make a view of a view."
             )
+            raise ValueError(msg)
         self._is_view = True
-        if isinstance(oidx, (int, np.integer)):
+        if isinstance(oidx, int | np.integer):
             if not (-adata_ref.n_obs <= oidx < adata_ref.n_obs):
-                raise IndexError(f"Observation index `{oidx}` is out of range.")
+                msg = f"Observation index `{oidx}` is out of range."
+                raise IndexError(msg)
             oidx += adata_ref.n_obs * (oidx < 0)
             oidx = slice(oidx, oidx + 1, 1)
-        if isinstance(vidx, (int, np.integer)):
+        if isinstance(vidx, int | np.integer):
             if not (-adata_ref.n_vars <= vidx < adata_ref.n_vars):
-                raise IndexError(f"Variable index `{vidx}` is out of range.")
+                msg = f"Variable index `{vidx}` is out of range."
+                raise IndexError(msg)
             vidx += adata_ref.n_vars * (vidx < 0)
             vidx = slice(vidx, vidx + 1, 1)
         if adata_ref.is_view:
@@ -297,12 +290,12 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         var_sub = adata_ref.var.iloc[vidx]
         # fix categories
         uns = copy(adata_ref._uns)
-        if settings.shall_remove_unused_categories:
+        if settings.remove_unused_categories:
             self._remove_unused_categories(adata_ref.obs, obs_sub, uns)
             self._remove_unused_categories(adata_ref.var, var_sub, uns)
         # set attributes
-        self._obs = DataFrameView(obs_sub, view_args=(self, "obs"))
-        self._var = DataFrameView(var_sub, view_args=(self, "var"))
+        self._obs = as_view(obs_sub, view_args=(self, "obs"))
+        self._var = as_view(var_sub, view_args=(self, "var"))
         self._uns = uns
 
         # set data
@@ -356,9 +349,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             # init from AnnData
             if isinstance(X, AnnData):
                 if any((obs, var, uns, obsm, varm, obsp, varp)):
-                    raise ValueError(
-                        "If `X` is a dict no further arguments must be provided."
-                    )
+                    msg = "If `X` is a dict no further arguments must be provided."
+                    raise ValueError(msg)
                 X, obs, var, uns, obsm, varm, obsp, varp, layers, raw = (
                     X._X,
                     X.obs,
@@ -393,7 +385,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         if X is not None:
             X = coerce_array(X, name="X")
             if shape is not None:
-                raise ValueError("`shape` needs to be `None` if `X` is not `None`.")
+                msg = "`shape` needs to be `None` if `X` is not `None`."
+                raise ValueError(msg)
             _check_2d_shape(X)
             # if type doesn’t match, a copy is made, otherwise, use a view
             if dtype is not None:
@@ -406,7 +399,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                     #       as in readwrite.read_10x_h5
                     if X.dtype != np.dtype(dtype):
                         X = X.astype(dtype)
-                elif isinstance(X, (ZarrArray, DaskArray)):
+                elif isinstance(X, ZarrArray | DaskArray):
                     X = X.astype(dtype)
                 else:  # is np.ndarray or a subclass, convert to true np.ndarray
                     X = np.asarray(X, dtype)
@@ -416,7 +409,11 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             source = "X"
         else:
             self._X = None
-            n_obs, n_vars = (None, None) if shape is None else shape
+            n_obs, n_vars = (
+                shape
+                if shape is not None
+                else _infer_shape(obs, var, obsm, varm, layers, obsp, varp)
+            )
             source = "shape"
 
         # annotations
@@ -433,7 +430,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             if isinstance(attr.index, pd.RangeIndex):
                 attr.index = idx
             elif not idx.equals(attr.index):
-                raise ValueError(f"Index of {attr_name} must match {x_name} of X.")
+                msg = f"Index of {attr_name} must match {x_name} of X."
+                raise ValueError(msg)
 
         # unstructured annotations
         self.uns = uns or OrderedDict()
@@ -447,13 +445,13 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         # Backwards compat for connectivities matrices in uns["neighbors"]
         _move_adj_mtx({"uns": self._uns, "obsp": self._obsp})
         self._check_dimensions()
-        if settings.shall_check_uniqueness:
+        if settings.check_uniqueness:
             self._check_uniqueness()
 
         if self.filename:
-            assert not isinstance(
-                raw, Raw
-            ), "got raw from other adata but also filename?"
+            assert not isinstance(raw, Raw), (
+                "got raw from other adata but also filename?"
+            )
             if {"raw", "raw.X"} & set(self.file):
                 raw = dict(X=None, **raw)
         if not raw:
@@ -469,7 +467,10 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         # layers
         self.layers = layers
 
-    def __sizeof__(self, show_stratified=None, with_disk: bool = False) -> int:
+    @old_positionals("show_stratified", "with_disk")
+    def __sizeof__(
+        self, *, show_stratified: bool = False, with_disk: bool = False
+    ) -> int:
         def get_size(X) -> int:
             def cs_to_bytes(X) -> int:
                 return int(X.data.nbytes + X.indptr.nbytes + X.indices.nbytes)
@@ -530,10 +531,11 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
     def __eq__(self, other):
         """Equality testing"""
-        raise NotImplementedError(
+        msg = (
             "Equality comparisons are not supported for AnnData objects, "
             "instead compare the desired attributes."
         )
+        raise NotImplementedError(msg)
 
     @property
     def shape(self) -> tuple[int, int]:
@@ -541,7 +543,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         return self.n_obs, self.n_vars
 
     @property
-    def X(self) -> ArrayDataStructureType | None:
+    def X(self) -> XDataType | None:
         """Data matrix of shape :attr:`n_obs` × :attr:`n_vars`."""
         if self.isbacked:
             if not self.file.is_open:
@@ -572,12 +574,11 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         #     return X
 
     @X.setter
-    def X(self, value: np.ndarray | sparse.spmatrix | SpArray | None):
+    def X(self, value: XDataType | None):
         if value is None:
             if self.isbacked:
-                raise NotImplementedError(
-                    "Cannot currently remove data matrix from backed object."
-                )
+                msg = "Cannot currently remove data matrix from backed object."
+                raise NotImplementedError(msg)
             if self.is_view:
                 self._init_as_actual(self.copy())
             self._X = None
@@ -631,7 +632,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                     if sparse.issparse(self._adata_ref._X) and isinstance(
                         value, np.ndarray
                     ):
-                        if isinstance(self._adata_ref.X, SpArray):
+                        if isinstance(self._adata_ref.X, CSArray):
                             memory_class = sparse.coo_array
                         else:
                             memory_class = sparse.coo_matrix
@@ -646,14 +647,17 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                             stacklevel=2,
                         )
                         value = value.toarray()
+                    warnings.warn(
+                        "Modifying `X` on a view results in data being overridden",
+                        ImplicitModificationWarning,
+                        stacklevel=2,
+                    )
                     self._adata_ref._X[oidx, vidx] = value
                 else:
                     self._X = value
         else:
-            raise ValueError(
-                f"Data matrix has wrong shape {value.shape}, "
-                f"need to be {self.shape}."
-            )
+            msg = f"Data matrix has wrong shape {value.shape}, need to be {self.shape}."
+            raise ValueError(msg)
 
     @X.deleter
     def X(self):
@@ -696,7 +700,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         The :attr:`raw` attribute is initialized with the current content
         of an object by setting::
 
-            adata.raw = adata
+            adata.raw = adata.copy()
 
         Its content can be deleted::
 
@@ -720,7 +724,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         if value is None:
             del self.raw
         elif not isinstance(value, AnnData):
-            raise ValueError("Can only init raw attribute with an AnnData object.")
+            msg = "Can only init raw attribute with an AnnData object."
+            raise ValueError(msg)
         else:
             if self.is_view:
                 self._init_as_actual(self.copy())
@@ -744,7 +749,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
     def _set_dim_df(self, value: pd.DataFrame, attr: Literal["obs", "var"]):
         if not isinstance(value, pd.DataFrame):
-            raise ValueError(f"Can only assign pd.DataFrame to {attr}.")
+            msg = f"Can only assign pd.DataFrame to {attr}."
+            raise ValueError(msg)
         raise_value_error_if_multiindex_columns(value, attr)
         value_idx = self._prep_dim_index(value.index, attr)
         if self.is_view:
@@ -760,19 +766,17 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         If a pd.Index is passed, this will use a reference, otherwise a new index object is created.
         """
         if self.shape[attr == "var"] != len(value):
-            raise ValueError(
-                f"Length of passed value for {attr}_names is {len(value)}, but this AnnData has shape: {self.shape}"
-            )
-        if isinstance(value, pd.Index) and not isinstance(
-            value.name, (str, type(None))
-        ):
-            raise ValueError(
+            msg = f"Length of passed value for {attr}_names is {len(value)}, but this AnnData has shape: {self.shape}"
+            raise ValueError(msg)
+        if isinstance(value, pd.Index) and not isinstance(value.name, str | type(None)):
+            msg = (
                 f"AnnData expects .{attr}.index.name to be a string or None, "
                 f"but you passed a name of type {type(value.name).__name__!r}"
             )
+            raise ValueError(msg)
         else:
             value = pd.Index(value)
-            if not isinstance(value.name, (str, type(None))):
+            if not isinstance(value.name, str | type(None)):
                 value.name = None
         if (
             len(value) > 0
@@ -857,9 +861,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
     @uns.setter
     def uns(self, value: MutableMapping):
         if not isinstance(value, MutableMapping):
-            raise ValueError(
-                "Only mutable mapping types (e.g. dict) are allowed for `.uns`."
-            )
+            msg = "Only mutable mapping types (e.g. dict) are allowed for `.uns`."
+            raise ValueError(msg)
         if isinstance(value, DictView):
             value = value.copy()
         if self.is_view:
@@ -961,7 +964,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         return self.file.filename
 
     @filename.setter
-    def filename(self, filename: PathLike | None):
+    def filename(self, filename: PathLike[str] | str | None):
         # convert early for later comparison
         filename = None if filename is None else Path(filename)
         # change from backing-mode back to full loading into memory
@@ -1023,8 +1026,10 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         oidx, vidx = self._normalize_indices(index)
         return AnnData(self, oidx=oidx, vidx=vidx, asview=True)
 
+    @staticmethod
+    @singledispatch
     def _remove_unused_categories(
-        self, df_full: pd.DataFrame, df_sub: pd.DataFrame, uns: dict[str, Any]
+        df_full: pd.DataFrame, df_sub: pd.DataFrame, uns: dict[str, Any]
     ):
         for k in df_full:
             if not isinstance(df_full[k].dtype, pd.CategoricalDtype):
@@ -1066,7 +1071,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
              New categories, the same number as the old categories.
         """
         if isinstance(categories, Mapping):
-            raise ValueError("Only list-like `categories` is supported.")
+            msg = "Only list-like `categories` is supported."
+            raise ValueError(msg)
         if key in self.obs:
             old_categories = self.obs[key].cat.categories.tolist()
             self.obs[key] = self.obs[key].cat.rename_categories(categories)
@@ -1074,7 +1080,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             old_categories = self.var[key].cat.categories.tolist()
             self.var[key] = self.var[key].cat.rename_categories(categories)
         else:
-            raise ValueError(f"{key} is neither in `.obs` nor in `.var`.")
+            msg = f"{key} is neither in `.obs` nor in `.var`."
+            raise ValueError(msg)
         # this is not a good solution
         # but depends on the scanpy conventions for storing the categorical key
         # as `groupby` in the `params` slot
@@ -1139,11 +1146,12 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 if not np.array_equal(c.categories, sorted_categories):
                     c = c.reorder_categories(sorted_categories)
                 if dont_modify:
-                    raise RuntimeError(
+                    msg = (
                         "Please call `.strings_to_categoricals()` on full "
                         "AnnData, not on this view. You might encounter this"
                         "error message while copying or writing to disk."
                     )
+                    raise RuntimeError(msg)
                 df[key] = c
                 logger.info(f"... storing {key!r} as categorical")
 
@@ -1170,11 +1178,10 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         self._init_as_actual(adata_subset)
 
     # TODO: Update, possibly remove
-    def __setitem__(
-        self, index: Index, val: int | float | np.ndarray | sparse.spmatrix
-    ):
+    def __setitem__(self, index: Index, val: float | XDataType):
         if self.is_view:
-            raise ValueError("Object is view and cannot be accessed with `[]`.")
+            msg = "Object is view and cannot be accessed with `[]`."
+            raise ValueError(msg)
         obs, var = self._normalize_indices(index)
         if not self.isbacked:
             self._X[obs, var] = val
@@ -1200,10 +1207,11 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         else:
             X = self.file["X"]
         if self.is_view:
-            raise ValueError(
+            msg = (
                 "You’re trying to transpose a view of an `AnnData`, "
                 "which is currently not implemented. Call `.copy()` before transposing."
             )
+            raise ValueError(msg)
 
         return AnnData(
             X=_safe_transpose(X) if X is not None else None,
@@ -1243,29 +1251,32 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         if layer is not None:
             X = self.layers[layer]
         elif not self._has_X():
-            raise ValueError("X is None, cannot convert to dataframe.")
+            msg = "X is None, cannot convert to dataframe."
+            raise ValueError(msg)
         else:
             X = self.X
         if issparse(X):
             X = X.toarray()
         return pd.DataFrame(X, index=self.obs_names, columns=self.var_names)
 
-    def _get_X(self, use_raw=False, layer=None):
+    def _get_X(self, *, use_raw: bool = False, layer: str | None = None):
         """\
         Convenience method for getting expression values
         with common arguments and error handling.
         """
         is_layer = layer is not None
         if use_raw and is_layer:
-            raise ValueError(
+            msg = (
                 "Cannot use expression from both layer and raw. You provided:"
                 f"`use_raw={use_raw}` and `layer={layer}`"
             )
+            raise ValueError(msg)
         if is_layer:
             return self.layers[layer]
         elif use_raw:
             if self.raw is None:
-                raise ValueError("This AnnData doesn’t have a value in `.raw`.")
+                msg = "This AnnData doesn’t have a value in `.raw`."
+                raise ValueError(msg)
             return self.raw.X
         else:
             return self.X
@@ -1334,8 +1345,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 layer = None
         return get_vector(self, k, "var", "obs", layer=layer)
 
-    @utils.deprecated("obs_vector")
-    def _get_obs_array(self, k, use_raw=False, layer=None):
+    @deprecated("obs_vector")
+    def _get_obs_array(self, k, use_raw=False, layer=None):  # noqa: FBT002
         """\
         Get an array from the layer (default layer='X') along the :attr:`obs`
         dimension by first looking up `obs.keys` and then :attr:`obs_names`.
@@ -1345,8 +1356,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         else:
             return self.raw.obs_vector(k)
 
-    @utils.deprecated("var_vector")
-    def _get_var_array(self, k, use_raw=False, layer=None):
+    @deprecated("var_vector")
+    def _get_var_array(self, k, use_raw=False, layer=None):  # noqa: FBT002
         """\
         Get an array from the layer (default layer='X') along the :attr:`var`
         dimension by first looking up `var.keys` and then :attr:`var_names`.
@@ -1360,10 +1371,11 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         """Creating AnnData with attributes optionally specified via kwargs."""
         if self.isbacked:
             if "X" not in kwargs or (self.raw is not None and "raw" not in kwargs):
-                raise NotImplementedError(
+                msg = (
                     "This function does not currently handle backed objects "
                     "internally, this should be dealt with before."
                 )
+                raise NotImplementedError(msg)
         new = {}
 
         for key in ["obs", "var", "obsm", "varm", "obsp", "varp", "layers"]:
@@ -1385,7 +1397,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             new["raw"] = self.raw.copy()
         return AnnData(**new)
 
-    def to_memory(self, copy=False) -> AnnData:
+    @old_positionals("copy")
+    def to_memory(self, *, copy: bool = False) -> AnnData:
         """Return a new AnnData object with all backed arrays loaded into memory.
 
         Params
@@ -1399,7 +1412,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         .. code:: python
 
             import anndata
-            backed = anndata.read_h5ad("file.h5ad", backed="r")
+            backed = anndata.io.read_h5ad("file.h5ad", backed="r")
             mem = backed[backed.obs["cluster"] == "a", :].to_memory()
         """
         new = {}
@@ -1416,13 +1429,13 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         ]:
             attr = getattr(self, attr_name, None)
             if attr is not None:
-                new[attr_name] = to_memory(attr, copy)
+                new[attr_name] = to_memory(attr, copy=copy)
 
         if self.raw is not None:
             new["raw"] = {
-                "X": to_memory(self.raw.X, copy),
-                "var": to_memory(self.raw.var, copy),
-                "varm": to_memory(self.raw.varm, copy),
+                "X": to_memory(self.raw.X, copy=copy),
+                "var": to_memory(self.raw.var, copy=copy),
+                "varm": to_memory(self.raw.varm, copy=copy),
             }
 
         if self.isbacked:
@@ -1430,7 +1443,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
         return AnnData(**new)
 
-    def copy(self, filename: PathLike | None = None) -> AnnData:
+    def copy(self, filename: PathLike[str] | str | None = None) -> AnnData:
         """Full copy, optionally on disk."""
         if not self.isbacked:
             if self.is_view and self._has_X():
@@ -1444,22 +1457,22 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             else:
                 return self._mutated_copy()
         else:
-            from .._io import read_h5ad, write_h5ad
+            from ..io import read_h5ad, write_h5ad
 
             if filename is None:
-                raise ValueError(
+                msg = (
                     "To copy an AnnData object in backed mode, "
                     "pass a filename: `.copy(filename='myfilename.h5ad')`. "
                     "To load the object into memory, use `.to_memory()`."
                 )
+                raise ValueError(msg)
             mode = self.file._filemode
             write_h5ad(filename, self)
             return read_h5ad(filename, backed=mode)
 
     @deprecated(
         "anndata.concat",
-        FutureWarning,
-        "See the tutorial for concat at: "
+        add_msg="See the tutorial for concat at: "
         "https://anndata.readthedocs.io/en/latest/concatenation.html",
         hide=False,
     )
@@ -1687,7 +1700,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         from .merge import concat, merge_dataframes, merge_outer, merge_same
 
         if self.isbacked:
-            raise ValueError("Currently, concatenate only works in memory mode.")
+            msg = "Currently, concatenate only works in memory mode."
+            raise ValueError(msg)
 
         if len(adatas) == 0:
             return self.copy()
@@ -1710,7 +1724,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         # Backwards compat (some of this could be more efficient)
         # obs used to always be an outer join
         sparse_class = sparse.csr_matrix
-        if any(isinstance(a.X, SpArray) for a in all_adatas):
+        if any(isinstance(a.X, CSArray) for a in all_adatas):
             sparse_class = sparse.csr_array
         out.obs = concat(
             [AnnData(sparse_class(a.shape), obs=a.obs) for a in all_adatas],
@@ -1761,9 +1775,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             utils.warn_names_duplicates("var")
 
     def __contains__(self, key: Any):
-        raise AttributeError(
-            "AnnData has no attribute __contains__, don’t check `in adata`."
-        )
+        msg = "AnnData has no attribute __contains__, don’t check `in adata`."
+        raise AttributeError(msg)
 
     def _check_dimensions(self, key=None):
         if key is None:
@@ -1775,23 +1788,28 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 not all([axis_len(o, 0) == self.n_obs for o in self.obsm.values()])
                 and len(self.obsm.dim_names) != self.n_obs
             ):
-                raise ValueError(
+                msg = (
                     "Observations annot. `obsm` must have number of rows of `X`"
                     f" ({self.n_obs}), but has {len(self.obsm)} rows."
                 )
+                raise ValueError(msg)
         if "varm" in key:
             if (
                 not all([axis_len(v, 0) == self.n_vars for v in self.varm.values()])
                 and len(self.varm.dim_names) != self.n_vars
             ):
-                raise ValueError(
+                msg = (
                     "Variables annot. `varm` must have number of columns of `X`"
                     f" ({self.n_vars}), but has {len(self.varm)} rows."
                 )
+                raise ValueError(msg)
 
+    @old_positionals("compression", "compression_opts", "as_dense")
     def write_h5ad(
         self,
-        filename: PathLike | None = None,
+        filename: PathLike[str] | str | None = None,
+        *,
+        convert_strings_to_categoricals: bool = True,
         compression: Literal["gzip", "lzf"] | None = None,
         compression_opts: int | Any = None,
         as_dense: Sequence[str] = (),
@@ -1815,6 +1833,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         ----------
         filename
             Filename of data file. Defaults to backing file.
+        convert_strings_to_categoricals
+            Convert string columns to categorical.
         compression
             For [`lzf`, `gzip`], see the h5py :ref:`dataset_compression`.
 
@@ -1858,16 +1878,18 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             Sparse arrays in AnnData object to write as dense. Currently only
             supports `X` and `raw/X`.
         """
-        from .._io import write_h5ad
+        from ..io import write_h5ad
 
         if filename is None and not self.isbacked:
-            raise ValueError("Provide a filename!")
+            msg = "Provide a filename!"
+            raise ValueError(msg)
         if filename is None:
             filename = self.filename
 
         write_h5ad(
             Path(filename),
             self,
+            convert_strings_to_categoricals=convert_strings_to_categoricals,
             compression=compression,
             compression_opts=compression_opts,
             as_dense=as_dense,
@@ -1878,7 +1900,10 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
     write = write_h5ad  # a shortcut and backwards compat
 
-    def write_csvs(self, dirname: PathLike, skip_data: bool = True, sep: str = ","):
+    @old_positionals("skip_data", "sep")
+    def write_csvs(
+        self, dirname: PathLike[str] | str, *, skip_data: bool = True, sep: str = ","
+    ):
         """\
         Write annotation to `.csv` files.
 
@@ -1894,11 +1919,14 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         sep
              Separator for the data.
         """
-        from .._io import write_csvs
+        from ..io import write_csvs
 
         write_csvs(dirname, self, skip_data=skip_data, sep=sep)
 
-    def write_loom(self, filename: PathLike, write_obsm_varm: bool = False):
+    @old_positionals("write_obsm_varm")
+    def write_loom(
+        self, filename: PathLike[str] | str, *, write_obsm_varm: bool = False
+    ):
         """\
         Write `.loom`-formatted hdf5 file.
 
@@ -1907,14 +1935,17 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         filename
             The filename.
         """
-        from .._io import write_loom
+        from ..io import write_loom
 
         write_loom(filename, self, write_obsm_varm=write_obsm_varm)
 
+    @old_positionals("chunks")
     def write_zarr(
         self,
-        store: MutableMapping | PathLike,
-        chunks: bool | int | tuple[int, ...] | None = None,
+        store: StoreLike,
+        *,
+        chunks: tuple[int, ...] | None = None,
+        convert_strings_to_categoricals: bool = True,
     ):
         """\
         Write a hierarchical Zarr array store.
@@ -1925,10 +1956,24 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             The filename, a :class:`~typing.MutableMapping`, or a Zarr storage class.
         chunks
             Chunk shape.
+        convert_strings_to_categoricals
+            Convert string columns to categorical.
         """
-        from .._io import write_zarr
+        from ..io import write_zarr
 
-        write_zarr(store, self, chunks=chunks)
+        # TODO: What is a bool for chunks supposed to do?
+        if isinstance(chunks, bool):
+            msg = (
+                "Passing `write_zarr(adata, chunks=True)` is no longer supported. "
+                "Please pass `write_zarr(adata)` instead."
+            )
+            raise ValueError(msg)
+        write_zarr(
+            store,
+            self,
+            chunks=chunks,
+            convert_strings_to_categoricals=convert_strings_to_categoricals,
+        )
 
     def chunked_X(self, chunk_size: int | None = None):
         """\
@@ -1951,9 +1996,11 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         if start < n:
             yield (self.X[start:n], start, n)
 
+    @old_positionals("replace")
     def chunk_X(
         self,
         select: int | Sequence[int] | np.ndarray = 1000,
+        *,
         replace: bool = True,
     ):
         """\
@@ -1976,10 +2023,11 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         if isinstance(select, int):
             select = select if select < self.n_obs else self.n_obs
             choice = np.random.choice(self.n_obs, select, replace)
-        elif isinstance(select, (np.ndarray, Sequence)):
+        elif isinstance(select, np.ndarray | Sequence):
             choice = np.asarray(select)
         else:
-            raise ValueError("select should be int or array")
+            msg = "select should be int or array"
+            raise ValueError(msg)
 
         reverse = None
         if self.isbacked:
@@ -2011,7 +2059,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
     # --------------------------------------------------------------------------
 
     @property
-    @utils.deprecated("is_view")
+    @deprecated("is_view")
     def isview(self):
         return self.is_view
 
@@ -2041,3 +2089,49 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         values = getattr(self, a)[keys].values
         getattr(self, a).drop(keys, axis=1, inplace=True)
         return values
+
+
+def _check_2d_shape(X):
+    """\
+    Check shape of array or sparse matrix.
+
+    Assure that X is always 2D: Unlike numpy we always deal with 2D arrays.
+    """
+    if X.dtype.names is None and len(X.shape) != 2:
+        msg = f"X needs to be 2-dimensional, not {len(X.shape)}-dimensional."
+        raise ValueError(msg)
+
+
+def _infer_shape_for_axis(
+    xxx: pd.DataFrame | Mapping[str, Iterable[Any]] | None,
+    xxxm: np.ndarray | Mapping[str, Sequence[Any]] | None,
+    layers: Mapping[str, np.ndarray | sparse.spmatrix] | None,
+    xxxp: np.ndarray | Mapping[str, Sequence[Any]] | None,
+    axis: Literal[0, 1],
+) -> int | None:
+    for elem in [xxx, xxxm, xxxp]:
+        if elem is not None and hasattr(elem, "shape"):
+            return elem.shape[0]
+    for elem, id in zip([layers, xxxm, xxxp], ["layers", "xxxm", "xxxp"]):
+        if elem is not None:
+            elem = cast("Mapping", elem)
+            for sub_elem in elem.values():
+                if hasattr(sub_elem, "shape"):
+                    size = cast("int", sub_elem.shape[axis if id == "layers" else 0])
+                    return size
+    return None
+
+
+def _infer_shape(
+    obs: pd.DataFrame | Mapping[str, Iterable[Any]] | None = None,
+    var: pd.DataFrame | Mapping[str, Iterable[Any]] | None = None,
+    obsm: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
+    varm: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
+    layers: Mapping[str, np.ndarray | sparse.spmatrix] | None = None,
+    obsp: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
+    varp: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
+):
+    return (
+        _infer_shape_for_axis(obs, obsm, layers, obsp, 0),
+        _infer_shape_for_axis(var, varm, layers, varp, 1),
+    )
