@@ -4,42 +4,46 @@ Code for merging/ concatenating AnnData objects.
 
 from __future__ import annotations
 
-import typing
 from collections import OrderedDict
 from collections.abc import Callable, Mapping, MutableSet
 from functools import partial, reduce, singledispatch
 from itertools import repeat
 from operator import and_, or_, sub
-from typing import Literal, TypeVar
+from typing import TYPE_CHECKING, Literal, TypeVar
 from warnings import warn
 
 import numpy as np
 import pandas as pd
+import scipy
 from natsort import natsorted
+from packaging.version import Version
 from scipy import sparse
-from scipy.sparse import spmatrix
 
+from anndata._core.file_backing import to_memory
 from anndata._warnings import ExperimentalFeatureWarning
 
 from ..compat import (
-    CAN_USE_SPARSE_ARRAY,
     AwkArray,
+    CSArray,
+    CSMatrix,
     CupyArray,
     CupyCSRMatrix,
     CupySparseMatrix,
     DaskArray,
-    SpArray,
     _map_cat_to_str,
 )
 from ..utils import asarray, axis_len, warn_once
 from .anndata import AnnData
 from .index import _subset, make_slice
 
-if typing.TYPE_CHECKING:
-    from collections.abc import Collection, Iterable, Sequence
+if TYPE_CHECKING:
+    from collections.abc import Collection, Generator, Iterable, Sequence
     from typing import Any
 
     from pandas.api.extensions import ExtensionDtype
+
+    from anndata._types import Join_T
+    from anndata.experimental.backed._compat import DataArray, Dataset2D
 
 T = TypeVar("T")
 
@@ -135,7 +139,7 @@ def equal_dask_array(a, b) -> bool:
     if isinstance(b, DaskArray):
         if tokenize(a) == tokenize(b):
             return True
-    if isinstance(a._meta, spmatrix):
+    if isinstance(a._meta, CSMatrix):
         # TODO: Maybe also do this in the other case?
         return da.map_blocks(equal, a, b, drop_axis=(0, 1)).all()
     else:
@@ -165,8 +169,8 @@ def equal_series(a, b) -> bool:
     return a.equals(b)
 
 
-@equal.register(sparse.spmatrix)
-@equal.register(SpArray)
+@equal.register(CSMatrix)
+@equal.register(CSArray)
 @equal.register(CupySparseMatrix)
 def equal_sparse(a, b) -> bool:
     # It's a weird api, don't blame me
@@ -174,7 +178,7 @@ def equal_sparse(a, b) -> bool:
 
     xp = array_api_compat.array_namespace(a.data)
 
-    if isinstance(b, (CupySparseMatrix, sparse.spmatrix, SpArray)):
+    if isinstance(b, CupySparseMatrix | CSMatrix | CSArray):
         if isinstance(a, CupySparseMatrix):
             # Comparison broken for CSC matrices
             # https://github.com/cupy/cupy/issues/7757
@@ -205,13 +209,19 @@ def equal_awkward(a, b) -> bool:
     return ak.almost_equal(a, b)
 
 
-def as_sparse(x, use_sparse_array=False):
-    if not isinstance(x, (sparse.spmatrix, SpArray)):
-        if CAN_USE_SPARSE_ARRAY and use_sparse_array:
-            return sparse.csr_array(x)
-        return sparse.csr_matrix(x)
-    else:
-        return x
+def as_sparse(x, *, use_sparse_array: bool = False) -> CSMatrix | CSArray:
+    if not isinstance(x, CSMatrix | CSArray):
+        in_memory_array_class = (
+            sparse.csr_array if use_sparse_array else sparse.csr_matrix
+        )
+        if isinstance(x, DaskArray):
+            x = x.map_blocks(
+                sparse.csr_matrix,
+                meta=sparse.csr_matrix(x._meta),
+                dtype=x.dtype,
+            ).compute()
+        return in_memory_array_class(x)
+    return x
 
 
 def as_cp_sparse(x) -> CupySparseMatrix:
@@ -225,7 +235,9 @@ def as_cp_sparse(x) -> CupySparseMatrix:
         return cpsparse.csr_matrix(x)
 
 
-def unify_dtypes(dfs: Iterable[pd.DataFrame]) -> list[pd.DataFrame]:
+def unify_dtypes(
+    dfs: Iterable[pd.DataFrame | Dataset2D],
+) -> list[pd.DataFrame | Dataset2D]:
     """
     Attempts to unify datatypes from multiple dataframes.
 
@@ -302,7 +314,7 @@ def try_unifying_dtype(
         return None
 
 
-def check_combinable_cols(cols: list[pd.Index], join: Literal["inner", "outer"]):
+def check_combinable_cols(cols: list[pd.Index], join: Join_T):
     """Given columns for a set of dataframes, checks if the can be combined.
 
     Looks for if there are duplicated column names that would show up in the result.
@@ -318,11 +330,12 @@ def check_combinable_cols(cols: list[pd.Index], join: Literal["inner", "outer"])
 
     if len(problem_cols) > 0:
         problem_cols = list(problem_cols)
-        raise pd.errors.InvalidIndexError(
+        msg = (
             f"Cannot combine dataframes as some contained duplicated column names - "
             "causing ambiguity.\n\n"
             f"The problem columns are: {problem_cols}"
         )
+        raise pd.errors.InvalidIndexError(msg)
 
 
 # TODO: open PR or feature request to cupy
@@ -536,7 +549,7 @@ class Reindexer:
             return el
         if isinstance(el, pd.DataFrame):
             return self._apply_to_df(el, axis=axis, fill_value=fill_value)
-        elif isinstance(el, (sparse.spmatrix, SpArray, CupySparseMatrix)):
+        elif isinstance(el, CSMatrix | CSArray | CupySparseMatrix):
             return self._apply_to_sparse(el, axis=axis, fill_value=fill_value)
         elif isinstance(el, AwkArray):
             return self._apply_to_awkward(el, axis=axis, fill_value=fill_value)
@@ -563,7 +576,7 @@ class Reindexer:
             shape[axis] = len(self.new_idx)
             return da.broadcast_to(fill_value, tuple(shape))
 
-        indexer = self.old_idx.get_indexer(self.new_idx)
+        indexer = self.idx
         sub_el = _subset(el, make_slice(indexer, axis, len(shape)))
 
         if any(indexer == -1):
@@ -606,7 +619,7 @@ class Reindexer:
             shape[axis] = len(self.new_idx)
             return np.broadcast_to(fill_value, tuple(shape))
 
-        indexer = self.old_idx.get_indexer(self.new_idx)
+        indexer = self.idx
 
         # Indexes real fast, and does outer indexing
         return pd.api.extensions.take(
@@ -614,8 +627,8 @@ class Reindexer:
         )
 
     def _apply_to_sparse(
-        self, el: sparse.spmatrix | SpArray, *, axis, fill_value=None
-    ) -> spmatrix:
+        self, el: CSMatrix | CSArray, *, axis, fill_value=None
+    ) -> CSMatrix:
         if isinstance(el, CupySparseMatrix):
             from cupyx.scipy import sparse
         else:
@@ -637,7 +650,7 @@ class Reindexer:
             shape[axis] = len(self.new_idx)
             shape = tuple(shape)
             if fill_value == 0:
-                if isinstance(el, SpArray):
+                if isinstance(el, CSArray):
                     memory_class = sparse.csr_array
                 else:
                     memory_class = sparse.csr_matrix
@@ -651,7 +664,7 @@ class Reindexer:
             idxmtx_dtype = xp.promote_types(el.dtype, xp.array(fill_value).dtype)
         else:
             idxmtx_dtype = bool
-        if isinstance(el, SpArray):
+        if isinstance(el, CSArray):
             memory_class = sparse.coo_array
         else:
             memory_class = sparse.coo_matrix
@@ -699,16 +712,19 @@ class Reindexer:
                 return el[self.new_idx]
             else:  # outer join
                 # TODO: this code isn't actually hit, we should refactor
-                raise Exception("This should be unreachable, please open an issue.")
+                msg = "This should be unreachable, please open an issue."
+                raise Exception(msg)
         else:
             if len(self.new_idx) > len(self.old_idx):
                 el = ak.pad_none(el, 1, axis=axis)  # axis == 0
-            return el[self.old_idx.get_indexer(self.new_idx)]
+            return el[self.idx]
+
+    @property
+    def idx(self):
+        return self.old_idx.get_indexer(self.new_idx)
 
 
-def merge_indices(
-    inds: Iterable[pd.Index], join: Literal["inner", "outer"]
-) -> pd.Index:
+def merge_indices(inds: Iterable[pd.Index], join: Join_T) -> pd.Index:
     if join == "inner":
         return reduce(lambda x, y: x.intersection(y), inds)
     elif join == "outer":
@@ -723,7 +739,11 @@ def default_fill_value(els):
 
     This is largely due to backwards compat, and might not be the ideal solution.
     """
-    if any(isinstance(el, (sparse.spmatrix, SpArray)) for el in els):
+    if any(
+        isinstance(el, CSMatrix | CSArray)
+        or (isinstance(el, DaskArray) and isinstance(el._meta, CSMatrix | CSArray))
+        for el in els
+    ):
         return 0
     else:
         return np.nan
@@ -758,19 +778,28 @@ def np_bool_to_pd_bool_array(df: pd.DataFrame):
 
 
 def concat_arrays(arrays, reindexers, axis=0, index=None, fill_value=None):
+    from anndata.experimental.backed._compat import Dataset2D
+
     arrays = list(arrays)
     if fill_value is None:
         fill_value = default_fill_value(arrays)
 
+    if any(isinstance(a, Dataset2D) for a in arrays):
+        if any(isinstance(a, pd.DataFrame) for a in arrays):
+            arrays = [to_memory(a) if isinstance(a, Dataset2D) else a for a in arrays]
+        elif not all(isinstance(a, Dataset2D) for a in arrays):
+            msg = f"Cannot concatenate a Dataset2D with other array types {[type(a) for a in arrays if not isinstance(a, Dataset2D)]}."
+            raise ValueError(msg)
+        else:
+            return concat_dataset2d_on_annot_axis(arrays, join="outer")
     if any(isinstance(a, pd.DataFrame) for a in arrays):
         # TODO: This is hacky, 0 is a sentinel for outer_concat_aligned_mapping
         if not all(
             isinstance(a, pd.DataFrame) or a is MissingVal or 0 in a.shape
             for a in arrays
         ):
-            raise NotImplementedError(
-                "Cannot concatenate a dataframe with other array types."
-            )
+            msg = "Cannot concatenate a dataframe with other array types."
+            raise NotImplementedError(msg)
         # TODO: behaviour here should be chosen through a merge strategy
         df = pd.concat(
             unify_dtypes(f(x) for f, x in zip(reindexers, arrays)),
@@ -785,20 +814,18 @@ def concat_arrays(arrays, reindexers, axis=0, index=None, fill_value=None):
         if not all(
             isinstance(a, AwkArray) or a is MissingVal or 0 in a.shape for a in arrays
         ):
-            raise NotImplementedError(
-                "Cannot concatenate an AwkwardArray with other array types."
-            )
+            msg = "Cannot concatenate an AwkwardArray with other array types."
+            raise NotImplementedError(msg)
 
         return ak.concatenate([f(a) for f, a in zip(reindexers, arrays)], axis=axis)
     elif any(isinstance(a, CupySparseMatrix) for a in arrays):
         import cupyx.scipy.sparse as cpsparse
 
         if not all(
-            isinstance(a, (CupySparseMatrix, CupyArray)) or 0 in a.shape for a in arrays
+            isinstance(a, CupySparseMatrix | CupyArray) or 0 in a.shape for a in arrays
         ):
-            raise NotImplementedError(
-                "Cannot concatenate a cupy array with other array types."
-            )
+            msg = "Cannot concatenate a cupy array with other array types."
+            raise NotImplementedError(msg)
         sparse_stack = (cpsparse.vstack, cpsparse.hstack)[axis]
         return sparse_stack(
             [
@@ -811,9 +838,8 @@ def concat_arrays(arrays, reindexers, axis=0, index=None, fill_value=None):
         import cupy as cp
 
         if not all(isinstance(a, CupyArray) or 0 in a.shape for a in arrays):
-            raise NotImplementedError(
-                "Cannot concatenate a cupy array with other array types."
-            )
+            msg = "Cannot concatenate a cupy array with other array types."
+            raise NotImplementedError(msg)
         return cp.concatenate(
             [
                 f(cp.asarray(x), fill_value=fill_value, axis=1 - axis)
@@ -821,10 +847,10 @@ def concat_arrays(arrays, reindexers, axis=0, index=None, fill_value=None):
             ],
             axis=axis,
         )
-    elif any(isinstance(a, (sparse.spmatrix, SpArray)) for a in arrays):
+    elif any(isinstance(a, CSMatrix | CSArray) for a in arrays):
         sparse_stack = (sparse.vstack, sparse.hstack)[axis]
-        use_sparse_array = any(issubclass(type(a), SpArray) for a in arrays)
-        return sparse_stack(
+        use_sparse_array = any(issubclass(type(a), CSArray) for a in arrays)
+        mat = sparse_stack(
             [
                 f(
                     as_sparse(a, use_sparse_array=use_sparse_array),
@@ -835,6 +861,13 @@ def concat_arrays(arrays, reindexers, axis=0, index=None, fill_value=None):
             ],
             format="csr",
         )
+        scipy_version = Version(scipy.__version__)
+        # Bug where xstack produces a matrix not an array in 1.11.*
+        if use_sparse_array and (scipy_version.major, scipy_version.minor) == (1, 11):
+            if mat.format == "csc":
+                return sparse.csc_array(mat)
+            return sparse.csr_array(mat)
+        return mat
     else:
         return np.concatenate(
             [
@@ -845,17 +878,23 @@ def concat_arrays(arrays, reindexers, axis=0, index=None, fill_value=None):
         )
 
 
-def inner_concat_aligned_mapping(mappings, *, reindexers=None, index=None, axis=0):
+def inner_concat_aligned_mapping(
+    mappings, *, reindexers=None, index=None, axis=0, concat_axis=None
+):
+    if concat_axis is None:
+        concat_axis = axis
     result = {}
 
     for k in intersect_keys(mappings):
         els = [m[k] for m in mappings]
         if reindexers is None:
-            cur_reindexers = gen_inner_reindexers(els, new_index=index, axis=axis)
+            cur_reindexers = gen_inner_reindexers(
+                els, new_index=index, axis=concat_axis
+            )
         else:
             cur_reindexers = reindexers
 
-        result[k] = concat_arrays(els, cur_reindexers, index=index, axis=axis)
+        result[k] = concat_arrays(els, cur_reindexers, index=index, axis=concat_axis)
     return result
 
 
@@ -873,9 +912,8 @@ def gen_inner_reindexers(els, new_index, axis: Literal[0, 1] = 0):
         reindexers = [Reindexer(df_indices(el), common_ind) for el in els]
     elif any(isinstance(el, AwkArray) for el in els if not_missing(el)):
         if not all(isinstance(el, AwkArray) for el in els if not_missing(el)):
-            raise NotImplementedError(
-                "Cannot concatenate an AwkwardArray with other array types."
-            )
+            msg = "Cannot concatenate an AwkwardArray with other array types."
+            raise NotImplementedError(msg)
         common_keys = intersect_keys(el.fields for el in els)
         reindexers = [
             Reindexer(pd.Index(el.fields), pd.Index(list(common_keys))) for el in els
@@ -901,9 +939,8 @@ def gen_outer_reindexers(els, shapes, new_index: pd.Index, *, axis=0):
         import awkward as ak
 
         if not all(isinstance(el, AwkArray) for el in els if not_missing(el)):
-            raise NotImplementedError(
-                "Cannot concatenate an AwkwardArray with other array types."
-            )
+            msg = "Cannot concatenate an AwkwardArray with other array types."
+            raise NotImplementedError(msg)
         warn_once(
             "Outer joins on awkward.Arrays will have different return values in the future. "
             "For details, and to offer input, please see:\n\n\t"
@@ -932,35 +969,67 @@ def gen_outer_reindexers(els, shapes, new_index: pd.Index, *, axis=0):
     return reindexers
 
 
+def missing_element(
+    n: int,
+    els: list[CSArray | CSMatrix | np.ndarray | DaskArray],
+    axis: Literal[0, 1] = 0,
+    fill_value: Any | None = None,
+    off_axis_size: int = 0,
+) -> np.ndarray | DaskArray:
+    """Generates value to use when there is a missing element."""
+    should_return_dask = any(isinstance(el, DaskArray) for el in els)
+    # 0 sized array for in-memory prevents allocating unnecessary memory while preserving broadcasting.
+    shape = (n, off_axis_size) if axis == 0 else (off_axis_size, n)
+    if should_return_dask:
+        import dask.array as da
+
+        return da.full(
+            shape, default_fill_value(els) if fill_value is None else fill_value
+        )
+    return np.zeros(shape, dtype=bool)
+
+
 def outer_concat_aligned_mapping(
-    mappings, *, reindexers=None, index=None, axis=0, fill_value=None
+    mappings, *, reindexers=None, index=None, axis=0, concat_axis=None, fill_value=None
 ):
+    if concat_axis is None:
+        concat_axis = axis
     result = {}
     ns = [m.parent.shape[axis] for m in mappings]
-
-    def missing_element(n: int, axis: Literal[0, 1] = 0) -> np.ndarray:
-        """Generates value to use when there is a missing element."""
-        if axis == 0:
-            return np.zeros((n, 0), dtype=bool)
-        else:
-            return np.zeros((0, n), dtype=bool)
 
     for k in union_keys(mappings):
         els = [m.get(k, MissingVal) for m in mappings]
         if reindexers is None:
-            cur_reindexers = gen_outer_reindexers(els, ns, new_index=index, axis=axis)
+            cur_reindexers = gen_outer_reindexers(
+                els, ns, new_index=index, axis=concat_axis
+            )
         else:
             cur_reindexers = reindexers
 
+        # Dask needs to create a full array and can't do the size-0 trick
+        off_axis_size = 0
+        if any(isinstance(e, DaskArray) for e in els):
+            if not isinstance(cur_reindexers[0], Reindexer):  # pragma: no cover
+                msg = "Cannot re-index a dask array without a Reindexer"
+                raise ValueError(msg)
+            off_axis_size = cur_reindexers[0].idx.shape[0]
         # Handling of missing values here is hacky for dataframes
         # We should probably just handle missing elements for all types
         result[k] = concat_arrays(
             [
-                el if not_missing(el) else missing_element(n, axis=axis)
+                el
+                if not_missing(el)
+                else missing_element(
+                    n,
+                    axis=concat_axis,
+                    els=els,
+                    fill_value=fill_value,
+                    off_axis_size=off_axis_size,
+                )
                 for el, n in zip(els, ns)
             ],
             cur_reindexers,
-            axis=axis,
+            axis=concat_axis,
             index=index,
             fill_value=fill_value,
         )
@@ -971,7 +1040,7 @@ def concat_pairwise_mapping(
     mappings: Collection[Mapping], shapes: Collection[int], join_keys=intersect_keys
 ):
     result = {}
-    if any(any(isinstance(v, SpArray) for v in m.values()) for m in mappings):
+    if any(any(isinstance(v, CSArray) for v in m.values()) for m in mappings):
         sparse_class = sparse.csr_array
     else:
         sparse_class = sparse.csr_matrix
@@ -980,7 +1049,7 @@ def concat_pairwise_mapping(
         els = [
             m.get(k, sparse_class((s, s), dtype=bool)) for m, s in zip(mappings, shapes)
         ]
-        if all(isinstance(el, (CupySparseMatrix, CupyArray)) for el in els):
+        if all(isinstance(el, CupySparseMatrix | CupyArray) for el in els):
             result[k] = _cp_block_diag(els, format="csr")
         elif all(isinstance(el, DaskArray) for el in els):
             result[k] = _dask_block_diag(els)
@@ -1021,7 +1090,8 @@ def _resolve_axis(
         return (0, "obs")
     if axis in {1, "var"}:
         return (1, "var")
-    raise ValueError(f"`axis` must be either 0, 1, 'obs', or 'var', was {axis}")
+    msg = f"`axis` must be either 0, 1, 'obs', or 'var', was {axis}"
+    raise ValueError(msg)
 
 
 def axis_indices(adata: AnnData, axis: Literal["obs", 0, "var", 1]) -> pd.Index:
@@ -1044,20 +1114,194 @@ def concat_Xs(adatas, reindexers, axis, fill_value):
     if all(X is None for X in Xs):
         return None
     elif any(X is None for X in Xs):
-        raise NotImplementedError(
+        msg = (
             "Some (but not all) of the AnnData's to be concatenated had no .X value. "
             "Concatenation is currently only implemented for cases where all or none of"
             " the AnnData's have .X assigned."
         )
+        raise NotImplementedError(msg)
     else:
         return concat_arrays(Xs, reindexers, axis=axis, fill_value=fill_value)
 
 
+def make_dask_col_from_extension_dtype(
+    col: DataArray, *, use_only_object_dtype: bool = False
+) -> DaskArray:
+    """
+    Creates dask arrays from :class:`pandas.api.extensions.ExtensionArray` dtype :class:`xarray.DataArray`s.
+
+    Parameters
+    ----------
+    col
+        The columns to be converted
+    use_only_object_dtype
+        Whether or not to cast all :class:`pandas.api.extensions.ExtensionArray` dtypes to `object` type, by default False
+
+    Returns
+    -------
+    A :class:`dask.Array`: representation of the column.
+    """
+    import dask.array as da
+
+    from anndata._io.specs.lazy_methods import (
+        compute_chunk_layout_for_axis_size,
+        get_chunksize,
+        maybe_open_h5,
+    )
+    from anndata.experimental import read_elem_lazy
+    from anndata.experimental.backed._compat import DataArray
+    from anndata.experimental.backed._compat import xarray as xr
+
+    base_path_or_zarr_group = col.attrs.get("base_path_or_zarr_group")
+    elem_name = col.attrs.get("elem_name")
+    dims = col.dims
+    coords = col.coords.copy()
+    with maybe_open_h5(base_path_or_zarr_group, elem_name) as f:
+        maybe_chunk_size = get_chunksize(read_elem_lazy(f))
+        chunk_size = (
+            compute_chunk_layout_for_axis_size(
+                1000 if maybe_chunk_size is None else maybe_chunk_size[0], col.shape[0]
+            ),
+        )
+
+    def get_chunk(block_info=None):
+        # reopening is important to get around h5py's unserializable lock in processes
+        with maybe_open_h5(base_path_or_zarr_group, elem_name) as f:
+            v = read_elem_lazy(f)
+            variable = xr.Variable(
+                data=xr.core.indexing.LazilyIndexedArray(v), dims=dims
+            )
+            data_array = DataArray(
+                variable,
+                coords=coords,
+                dims=dims,
+            )
+            idx = tuple(
+                slice(start, stop) for start, stop in block_info[None]["array-location"]
+            )
+            chunk = np.array(data_array.data[idx].array)
+        return chunk
+
+    if col.dtype == "category" or col.dtype == "string" or use_only_object_dtype:
+        dtype = "object"
+    else:
+        dtype = col.dtype.numpy_dtype
+    return da.map_blocks(
+        get_chunk,
+        chunks=chunk_size,
+        meta=np.array([], dtype=dtype),
+        dtype=dtype,
+    )
+
+
+def make_xarray_extension_dtypes_dask(
+    annotations: Iterable[Dataset2D], *, use_only_object_dtype: bool = False
+) -> Generator[Dataset2D, None, None]:
+    """
+    Creates a generator of Dataset2D objects with dask arrays in place of :class:`pandas.api.extensions.ExtensionArray` dtype columns.
+
+    Parameters
+    ----------
+    annotations
+        The datasets to be altered
+    use_only_object_dtype
+        Whether or not to cast all :class:`pandas.api.extensions.ExtensionArray` dtypes to `object` type, by default False
+
+    Yields
+    ------
+    An altered dataset.
+    """
+    for a in annotations:
+        extension_cols = {
+            col for col in a.columns if pd.api.types.is_extension_array_dtype(a[col])
+        }
+
+        yield a.copy(
+            data={
+                name: (
+                    make_dask_col_from_extension_dtype(
+                        col, use_only_object_dtype=use_only_object_dtype
+                    )
+                    if name in extension_cols
+                    else col
+                )
+                for name, col in a.items()
+            }
+        )
+
+
+DS_CONCAT_DUMMY_INDEX_NAME = "concat_index"
+
+
+def concat_dataset2d_on_annot_axis(
+    annotations: Iterable[Dataset2D],
+    join: Join_T,
+) -> Dataset2D:
+    """Create a concatenate dataset from a list of :class:`~anndata.experimental.backed._xarray.Dataset2D` objects.
+    The goal of this function is to mimic `pd.concat(..., ignore_index=True)` so has some complicated logic
+    for handling the "index" to ensure (a) nothing is loaded into memory and (b) the true index is always tracked.
+
+    Parameters
+    ----------
+    annotations
+        The :class:`~anndata.experimental.backed._xarray.Dataset2D` objects to be concatenated.
+    join
+        Type of join operation
+
+    Returns
+    -------
+    Concatenated :class:`~anndata.experimental.backed._xarray.Dataset2D`
+    """
+    from anndata._io.specs.lazy_methods import DUMMY_RANGE_INDEX_KEY
+    from anndata.experimental.backed._compat import Dataset2D
+    from anndata.experimental.backed._compat import xarray as xr
+
+    annotations_re_indexed = []
+    for a in make_xarray_extension_dtypes_dask(annotations):
+        old_key = list(a.coords.keys())[0]
+        # First create a dummy index
+        a.coords[DS_CONCAT_DUMMY_INDEX_NAME] = (
+            old_key,
+            pd.RangeIndex(a[a.attrs["indexing_key"]].shape[0]).astype("str"),
+        )
+        # Set all the dimensions to this new dummy index
+        a = a.swap_dims({old_key: DS_CONCAT_DUMMY_INDEX_NAME})
+        # Move the old coordinate into a variable
+        old_coord = a.coords[old_key]
+        del a.coords[old_key]
+        a[old_key] = old_coord
+        annotations_re_indexed.append(a)
+    # Concat along the dummy index
+    ds = Dataset2D(
+        xr.concat(annotations_re_indexed, join=join, dim=DS_CONCAT_DUMMY_INDEX_NAME),
+        attrs={"indexing_key": f"true_{DS_CONCAT_DUMMY_INDEX_NAME}"},
+    )
+    ds.coords[DS_CONCAT_DUMMY_INDEX_NAME] = pd.RangeIndex(
+        ds.coords[DS_CONCAT_DUMMY_INDEX_NAME].shape[0]
+    ).astype("str")
+    # Drop any lingering dimensions (swap doesn't delete)
+    ds = ds.drop_dims(d for d in ds.dims if d != DS_CONCAT_DUMMY_INDEX_NAME)
+    # Create a new true index and then delete the columns resulting from the concatenation for each index.
+    # This includes the dummy column (which is neither a dimension nor a true indexing column)
+    index = xr.concat(
+        [a[a.attrs["indexing_key"]] for a in annotations_re_indexed],
+        dim=DS_CONCAT_DUMMY_INDEX_NAME,
+    )
+    # prevent duplicate values
+    index.coords[DS_CONCAT_DUMMY_INDEX_NAME] = ds.coords[DS_CONCAT_DUMMY_INDEX_NAME]
+    ds[f"true_{DS_CONCAT_DUMMY_INDEX_NAME}"] = index
+    for key in set(a.attrs["indexing_key"] for a in annotations_re_indexed):
+        del ds[key]
+    if DUMMY_RANGE_INDEX_KEY in ds:
+        del ds[DUMMY_RANGE_INDEX_KEY]
+    return ds
+
+
 def concat(
-    adatas: Collection[AnnData] | typing.Mapping[str, AnnData],
+    adatas: Collection[AnnData] | Mapping[str, AnnData],
     *,
     axis: Literal["obs", 0, "var", 1] = "obs",
-    join: Literal["inner", "outer"] = "inner",
+    join: Join_T = "inner",
     merge: StrategiesLiteral | Callable | None = None,
     uns_merge: StrategiesLiteral | Callable | None = None,
     label: str | None = None,
@@ -1090,6 +1334,8 @@ def concat(
         * `"unique"`: Elements for which there is only one possible value.
         * `"first"`: The first element seen at each from each position.
         * `"only"`: Elements that show up in only one of the objects.
+
+        For :class:`xarray.Dataset` objects, we use their :func:`xarray.merge` with `override` to stay lazy.
     uns_merge
         How the elements of `.uns` are selected. Uses the same set of strategies as
         the `merge` argument, except applied recursively.
@@ -1254,16 +1500,21 @@ def concat(
     >>> dict(ad.concat([a, b, c], uns_merge="first").uns)
     {'a': 1, 'b': 2, 'c': {'c.a': 3, 'c.b': 4, 'c.c': 5}}
     """
+
+    from anndata.experimental.backed._compat import Dataset2D
+    from anndata.experimental.backed._compat import xarray as xr
+
     # Argument normalization
     merge = resolve_merge_strategy(merge)
     uns_merge = resolve_merge_strategy(uns_merge)
 
     if isinstance(adatas, Mapping):
         if keys is not None:
-            raise TypeError(
+            msg = (
                 "Cannot specify categories in both mapping keys and using `keys`. "
                 "Only specify this once."
             )
+            raise TypeError(msg)
         keys, adatas = list(adatas.keys()), list(adatas.values())
     else:
         adatas = list(adatas)
@@ -1299,19 +1550,53 @@ def concat(
 
     # Annotation for concatenation axis
     check_combinable_cols([getattr(a, axis_name).columns for a in adatas], join=join)
-    concat_annot = pd.concat(
-        unify_dtypes(getattr(a, axis_name) for a in adatas),
-        join=join,
-        ignore_index=True,
+    annotations = [getattr(a, axis_name) for a in adatas]
+    are_any_annotations_dataframes = any(
+        isinstance(a, pd.DataFrame) for a in annotations
     )
-    concat_annot.index = concat_indices
+    if are_any_annotations_dataframes:
+        annotations_in_memory = (
+            to_memory(a) if isinstance(a, Dataset2D) else a for a in annotations
+        )
+        concat_annot = pd.concat(
+            unify_dtypes(annotations_in_memory),
+            join=join,
+            ignore_index=True,
+        )
+        concat_annot.index = concat_indices
+    else:
+        concat_annot = concat_dataset2d_on_annot_axis(annotations, join)
+        concat_indices.name = DS_CONCAT_DUMMY_INDEX_NAME
     if label is not None:
         concat_annot[label] = label_col
 
     # Annotation for other axis
-    alt_annot = merge_dataframes(
-        [getattr(a, alt_axis_name) for a in adatas], alt_indices, merge
+    alt_annotations = [getattr(a, alt_axis_name) for a in adatas]
+    are_any_alt_annotations_dataframes = any(
+        isinstance(a, pd.DataFrame) for a in alt_annotations
     )
+    if are_any_alt_annotations_dataframes:
+        alt_annotations_in_memory = [
+            to_memory(a) if isinstance(a, Dataset2D) else a for a in alt_annotations
+        ]
+        alt_annot = merge_dataframes(alt_annotations_in_memory, alt_indices, merge)
+    else:
+        # TODO: figure out mapping of our merge to theirs instead of just taking first, although this appears to be
+        # the only "lazy" setting so I'm not sure we really want that.
+        # Because of xarray's merge upcasting, it's safest to simply assume that all dtypes are objects.
+        annotations_with_only_dask = list(
+            make_xarray_extension_dtypes_dask(
+                alt_annotations, use_only_object_dtype=True
+            )
+        )
+        annotations_with_only_dask = [
+            a.rename({a.attrs["indexing_key"]: "merge_index"})
+            for a in annotations_with_only_dask
+        ]
+        alt_annot = Dataset2D(
+            xr.merge(annotations_with_only_dask, join=join, compat="override"),
+            attrs={"indexing_key": "merge_index"},
+        )
 
     X = concat_Xs(adatas, reindexers, axis=axis, fill_value=fill_value)
 
@@ -1324,13 +1609,17 @@ def concat(
         )
         join_keys = union_keys
     else:
-        raise AssertionError(f"{join=} should have been validated above by pd.concat")
+        msg = f"{join=} should have been validated above by pd.concat"
+        raise AssertionError(msg)
 
     layers = concat_aligned_mapping(
         [a.layers for a in adatas], axis=axis, reindexers=reindexers
     )
     concat_mapping = concat_aligned_mapping(
-        [getattr(a, f"{axis_name}m") for a in adatas], index=concat_indices
+        [getattr(a, f"{axis_name}m") for a in adatas],
+        axis=axis,
+        concat_axis=0,
+        index=concat_indices,
     )
     if pairwise:
         concat_pairwise = concat_pairwise_mapping(
