@@ -16,33 +16,45 @@ import warnings
 from abc import ABC
 from collections.abc import Iterable
 from functools import cached_property
-from itertools import accumulate, chain
+from itertools import accumulate, chain, pairwise
 from math import floor
 from pathlib import Path
 from typing import TYPE_CHECKING, NamedTuple
 
 import h5py
 import numpy as np
+import scipy
 import scipy.sparse as ss
+from packaging.version import Version
 from scipy.sparse import _sparsetools
-
-try:
-    # Not really important, just for IDEs to be more helpful
-    from scipy.sparse._compressed import _cs_matrix
-except ImportError:
-    from scipy.sparse import spmatrix as _cs_matrix
 
 from .. import abc
 from .._settings import settings
-from ..compat import H5Group, SpArray, ZarrArray, ZarrGroup, _read_attr
+from ..compat import (
+    CSArray,
+    CSMatrix,
+    H5Group,
+    ZarrArray,
+    ZarrGroup,
+    _read_attr,
+    is_zarr_v2,
+)
 from .index import _fix_slice_bounds, _subset, unpack_index
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
     from typing import Literal
 
+    from scipy.sparse._compressed import _cs_matrix
+
     from .._types import GroupStorageType
-    from .index import Index
+    from ..compat import H5Array
+    from .index import Index, Index1D
+else:
+    from scipy.sparse import spmatrix as _cs_matrix
+
+
+SCIPY_1_15 = Version(scipy.__version__) >= Version("1.15rc0")
 
 
 class BackedFormat(NamedTuple):
@@ -63,19 +75,28 @@ class BackedSparseMatrix(_cs_matrix):
     indices: GroupStorageType
     indptr: np.ndarray
 
-    def copy(self) -> ss.csr_matrix | ss.csc_matrix:
+    def copy(self) -> CSMatrix:
         if isinstance(self.data, h5py.Dataset):
             return sparse_dataset(self.data.parent).to_memory()
         if isinstance(self.data, ZarrArray):
             import zarr
 
-            return sparse_dataset(
-                zarr.open(
+            if is_zarr_v2():
+                sparse_group = zarr.open(
                     store=self.data.store,
                     mode="r",
                     chunk_store=self.data.chunk_store,  # chunk_store is needed, not clear why
                 )[Path(self.data.path).parent]
-            ).to_memory()
+            else:
+                anndata_group = zarr.open_group(store=self.data.store, mode="r")
+                sparse_group = anndata_group[
+                    str(
+                        Path(str(self.data.store_path))
+                        .relative_to(str(anndata_group.store_path))
+                        .parent
+                    )
+                ]
+            return sparse_dataset(sparse_group).to_memory()
         return super().copy()
 
     def _set_many(self, i: Iterable[int], j: Iterable[int], x):
@@ -97,9 +118,8 @@ class BackedSparseMatrix(_cs_matrix):
             return
 
         else:
-            raise ValueError(
-                "You cannot change the sparsity structure of a SparseDataset."
-            )
+            msg = "You cannot change the sparsity structure of a SparseDataset."
+            raise ValueError(msg)
             # replace where possible
             # mask = offsets > -1
             # # offsets[mask]
@@ -251,30 +271,42 @@ def slice_as_int(s: slice, l: int) -> int:
 def get_compressed_vectors(
     x: BackedSparseMatrix, row_idxs: Iterable[int]
 ) -> tuple[Sequence, Sequence, Sequence]:
-    slices = [slice(*(x.indptr[i : i + 2])) for i in row_idxs]
-    data = np.concatenate([x.data[s] for s in slices])
-    indices = np.concatenate([x.indices[s] for s in slices])
-    indptr = list(accumulate(chain((0,), (s.stop - s.start for s in slices))))
+    indptr_slices = [slice(*(x.indptr[i : i + 2])) for i in row_idxs]
+    # HDF5 cannot handle out-of-order integer indexing
+    if isinstance(x.data, ZarrArray):
+        as_np_indptr = np.concatenate(
+            [np.arange(s.start, s.stop) for s in indptr_slices]
+        )
+        data = x.data[as_np_indptr]
+        indices = x.indices[as_np_indptr]
+    else:
+        data = np.concatenate([x.data[s] for s in indptr_slices])
+        indices = np.concatenate([x.indices[s] for s in indptr_slices])
+    indptr = list(accumulate(chain((0,), (s.stop - s.start for s in indptr_slices))))
     return data, indices, indptr
 
 
 def get_compressed_vectors_for_slices(
     x: BackedSparseMatrix, slices: Iterable[slice]
 ) -> tuple[Sequence, Sequence, Sequence]:
-    indptr_sels = [x.indptr[slice(s.start, s.stop + 1)] for s in slices]
-    data = np.concatenate([x.data[s[0] : s[-1]] for s in indptr_sels])
-    indices = np.concatenate([x.indices[s[0] : s[-1]] for s in indptr_sels])
+    indptr_indices = [x.indptr[slice(s.start, s.stop + 1)] for s in slices]
+    indptr_limits = [slice(i[0], i[-1]) for i in indptr_indices]
+    # HDF5 cannot handle out-of-order integer indexing
+    if isinstance(x.data, ZarrArray):
+        indptr_int = np.concatenate([np.arange(s.start, s.stop) for s in indptr_limits])
+        data = x.data[indptr_int]
+        indices = x.indices[indptr_int]
+    else:
+        data = np.concatenate([x.data[s] for s in indptr_limits])
+        indices = np.concatenate([x.indices[s] for s in indptr_limits])
     # Need to track the size of the gaps in the slices to each indptr subselection
-    total = indptr_sels[0][0]
-    offsets = [total]
-    for i, sel in enumerate(indptr_sels[1:]):
-        total = (sel[0] - indptr_sels[i][-1]) + total
-        offsets.append(total)
-    start_indptr = indptr_sels[0] - offsets[0]
+    gaps = (s1.start - s0.stop for s0, s1 in pairwise(indptr_limits))
+    offsets = accumulate(chain([indptr_limits[0].start], gaps))
+    start_indptr = indptr_indices[0] - next(offsets)
     if len(slices) < 2:  # there is only one slice so no need to concatenate
         return data, indices, start_indptr
     end_indptr = np.concatenate(
-        [s[1:] - offsets[i + 1] for i, s in enumerate(indptr_sels[1:])]
+        [s[1:] - o for s, o in zip(indptr_indices[1:], offsets)]
     )
     indptr = np.concatenate([start_indptr, end_indptr])
     return data, indices, indptr
@@ -312,11 +344,12 @@ def get_memory_class(
 ) -> type[_cs_matrix]:
     for fmt, _, memory_class in FORMATS:
         if format == fmt:
-            if use_sparray_in_io and issubclass(memory_class, SpArray):
+            if use_sparray_in_io and issubclass(memory_class, CSArray):
                 return memory_class
-            elif not use_sparray_in_io and issubclass(memory_class, ss.spmatrix):
+            elif not use_sparray_in_io and issubclass(memory_class, CSMatrix):
                 return memory_class
-    raise ValueError(f"Format string {format} is not supported.")
+    msg = f"Format string {format} is not supported."
+    raise ValueError(msg)
 
 
 def get_backed_class(
@@ -324,11 +357,12 @@ def get_backed_class(
 ) -> type[BackedSparseMatrix]:
     for fmt, backed_class, _ in FORMATS:
         if format == fmt:
-            if use_sparray_in_io and issubclass(backed_class, SpArray):
+            if use_sparray_in_io and issubclass(backed_class, CSArray):
                 return backed_class
-            elif not use_sparray_in_io and issubclass(backed_class, ss.spmatrix):
+            elif not use_sparray_in_io and issubclass(backed_class, CSMatrix):
                 return backed_class
-    raise ValueError(f"Format string {format} is not supported.")
+    msg = f"Format string {format} is not supported."
+    raise ValueError(msg)
 
 
 def _get_group_format(group: GroupStorageType) -> str:
@@ -342,13 +376,22 @@ def _get_group_format(group: GroupStorageType) -> str:
 
 
 # Check for the overridden few methods above in our BackedSparseMatrix subclasses
-def is_sparse_indexing_overridden(format: Literal["csr", "csc"], row, col):
+def is_sparse_indexing_overridden(
+    format: Literal["csr", "csc"], row: Index1D, col: Index1D
+):
     major_indexer, minor_indexer = (row, col) if format == "csr" else (col, row)
     return isinstance(minor_indexer, slice) and (
-        (isinstance(major_indexer, (int, np.integer)))
+        (isinstance(major_indexer, int | np.integer))
         or (isinstance(major_indexer, slice))
         or (isinstance(major_indexer, np.ndarray) and major_indexer.ndim == 1)
     )
+
+
+def validate_indices(
+    mtx: BackedSparseMatrix, indices: tuple[Index1D, Index1D]
+) -> tuple[Index1D, Index1D]:
+    res = mtx._validate_indices(indices)
+    return res[0] if SCIPY_1_15 else res
 
 
 class BaseCompressedSparseDataset(abc._AbstractCSDataset, ABC):
@@ -365,9 +408,8 @@ class BaseCompressedSparseDataset(abc._AbstractCSDataset, ABC):
 
     @group.setter
     def group(self, val):
-        raise AttributeError(
-            f"Do not reset group on a {type(self)} with {val}.  Instead use `sparse_dataset` to make a new class."
-        )
+        msg = f"Do not reset group on a {type(self)} with {val}.  Instead use `sparse_dataset` to make a new class."
+        raise AttributeError(msg)
 
     @property
     def backend(self) -> Literal["zarr", "hdf5"]:
@@ -377,12 +419,13 @@ class BaseCompressedSparseDataset(abc._AbstractCSDataset, ABC):
         elif isinstance(self.group, H5Group):
             return "hdf5"
         else:
-            raise ValueError(f"Unknown group type {type(self.group)}")
+            msg = f"Unknown group type {type(self.group)}"
+            raise ValueError(msg)
 
     @property
     def dtype(self) -> np.dtype:
         """The :class:`numpy.dtype` of the `data` attribute of the sparse matrix."""
-        return self.group["data"].dtype
+        return self._data.dtype
 
     @classmethod
     def _check_group_format(cls, group):
@@ -407,14 +450,12 @@ class BaseCompressedSparseDataset(abc._AbstractCSDataset, ABC):
         name = type(self).__name__.removeprefix("_")
         return f"{name}: backend {self.backend}, shape {self.shape}, data_dtype {self.dtype}"
 
-    def __getitem__(
-        self, index: Index | tuple[()]
-    ) -> float | ss.csr_matrix | ss.csc_matrix | SpArray:
+    def __getitem__(self, index: Index | tuple[()]) -> float | CSMatrix | CSArray:
         indices = self._normalize_index(index)
         row, col = indices
         mtx = self._to_backed()
-        row_sp_matrix_validated, col_sp_matrix_validated = mtx._validate_indices(
-            (row, col)
+        row_sp_matrix_validated, col_sp_matrix_validated = validate_indices(
+            mtx, indices
         )
 
         # Handle masked indexing along major axis
@@ -438,10 +479,10 @@ class BaseCompressedSparseDataset(abc._AbstractCSDataset, ABC):
         # Not sure what the performance is on that operation
         # Also need to check if memory format is not matrix
         mtx_fmt = get_memory_class(
-            self.format, use_sparray_in_io=settings.shall_use_sparse_array_on_read
+            self.format, use_sparray_in_io=settings.use_sparse_array_on_read
         )
-        must_convert_to_array = issubclass(mtx_fmt, SpArray) and not isinstance(
-            sub, SpArray
+        must_convert_to_array = issubclass(mtx_fmt, CSArray) and not isinstance(
+            sub, CSArray
         )
         if isinstance(sub, BackedSparseMatrix) or must_convert_to_array:
             return mtx_fmt(sub)
@@ -460,15 +501,15 @@ class BaseCompressedSparseDataset(abc._AbstractCSDataset, ABC):
 
     def __setitem__(self, index: Index | tuple[()], value) -> None:
         warnings.warn(
-            "__setitem__ will likely be removed in the near future. We do not recommend relying on its stability.",
-            PendingDeprecationWarning,
+            "__setitem__ for backed sparse will be removed in the next anndata release.",
+            FutureWarning,
         )
         row, col = self._normalize_index(index)
         mock_matrix = self._to_backed()
         mock_matrix[row, col] = value
 
     # TODO: split to other classes?
-    def append(self, sparse_matrix: ss.csr_matrix | ss.csc_matrix | SpArray) -> None:
+    def append(self, sparse_matrix: CSMatrix | CSArray) -> None:
         """Append an in-memory or on-disk sparse matrix to the current object's store.
 
         Parameters
@@ -494,42 +535,45 @@ class BaseCompressedSparseDataset(abc._AbstractCSDataset, ABC):
 
         # Check input
         if not ss.issparse(sparse_matrix):
-            raise NotImplementedError(
+            msg = (
                 "Currently, only sparse matrices of equivalent format can be "
                 "appended to a SparseDataset."
             )
+            raise NotImplementedError(msg)
         if self.format not in {"csr", "csc"}:
-            raise NotImplementedError(
-                f"The append method for format {self.format} " f"is not implemented."
-            )
+            msg = f"The append method for format {self.format} is not implemented."
+            raise NotImplementedError(msg)
         if self.format != sparse_matrix.format:
-            raise ValueError(
+            msg = (
                 f"Matrices must have same format. Currently are "
                 f"{self.format!r} and {sparse_matrix.format!r}"
             )
-        indptr_offset = len(self.group["indices"])
+            raise ValueError(msg)
+        [indptr_offset] = self.group["indices"].shape
         if self.group["indptr"].dtype == np.int32:
-            new_nnz = indptr_offset + len(sparse_matrix.indices)
+            new_nnz = indptr_offset + sparse_matrix.indices.shape[0]
             if new_nnz >= np.iinfo(np.int32).max:
-                raise OverflowError(
+                msg = (
                     "This array was written with a 32 bit intptr, but is now large "
                     "enough to require 64 bit values. Please recreate the array with "
                     "a 64 bit indptr."
                 )
+                raise OverflowError(msg)
 
         # shape
         if self.format == "csr":
-            assert (
-                shape[1] == sparse_matrix.shape[1]
-            ), "CSR matrices must have same size of dimension 1 to be appended."
+            assert shape[1] == sparse_matrix.shape[1], (
+                "CSR matrices must have same size of dimension 1 to be appended."
+            )
             new_shape = (shape[0] + sparse_matrix.shape[0], shape[1])
         elif self.format == "csc":
-            assert (
-                shape[0] == sparse_matrix.shape[0]
-            ), "CSC matrices must have same size of dimension 0 to be appended."
+            assert shape[0] == sparse_matrix.shape[0], (
+                "CSC matrices must have same size of dimension 0 to be appended."
+            )
             new_shape = (shape[0], shape[1] + sparse_matrix.shape[1])
         else:
-            raise AssertionError("We forgot to update this branching to a new format")
+            msg = "We forgot to update this branching to a new format"
+            raise AssertionError(msg)
         if "h5sparse_shape" in self.group.attrs:
             del self.group.attrs["h5sparse_shape"]
         self.group.attrs["shape"] = new_shape
@@ -538,8 +582,13 @@ class BaseCompressedSparseDataset(abc._AbstractCSDataset, ABC):
         data = self.group["data"]
         orig_data_size = data.shape[0]
         data.resize((orig_data_size + sparse_matrix.data.shape[0],))
-        data[orig_data_size:] = sparse_matrix.data
-
+        # see https://github.com/zarr-developers/zarr-python/discussions/2712 for why we need to read first
+        append_data = sparse_matrix.data
+        append_indices = sparse_matrix.indices
+        if isinstance(sparse_matrix.data, ZarrArray) and not is_zarr_v2():
+            data[orig_data_size:] = append_data[...]
+        else:
+            data[orig_data_size:] = append_data
         # indptr
         indptr = self.group["indptr"]
         orig_data_size = indptr.shape[0]
@@ -547,15 +596,19 @@ class BaseCompressedSparseDataset(abc._AbstractCSDataset, ABC):
         indptr[orig_data_size:] = (
             sparse_matrix.indptr[1:].astype(np.int64) + indptr_offset
         )
-        # Clear cached property
-        if hasattr(self, "indptr"):
-            del self._indptr
 
         # indices
+        if isinstance(sparse_matrix.data, ZarrArray) and not is_zarr_v2():
+            append_indices = append_indices[...]
         indices = self.group["indices"]
         orig_data_size = indices.shape[0]
         indices.resize((orig_data_size + sparse_matrix.indices.shape[0],))
-        indices[orig_data_size:] = sparse_matrix.indices
+        indices[orig_data_size:] = append_indices
+
+        # Clear cached property
+        for attr in ["_indptr", "_indices", "_data"]:
+            if hasattr(self, attr):
+                delattr(self, attr)
 
     @cached_property
     def _indptr(self) -> np.ndarray:
@@ -567,21 +620,35 @@ class BaseCompressedSparseDataset(abc._AbstractCSDataset, ABC):
         arr = self.group["indptr"][...]
         return arr
 
+    @cached_property
+    def _indices(self) -> H5Array | ZarrArray:
+        """\
+        Cache access to the indices to prevent unnecessary reads of the zarray
+        """
+        return self.group["indices"]
+
+    @cached_property
+    def _data(self) -> H5Array | ZarrArray:
+        """\
+        Cache access to the data to prevent unnecessary reads of the zarray
+        """
+        return self.group["data"]
+
     def _to_backed(self) -> BackedSparseMatrix:
         format_class = get_backed_class(self.format)
         mtx = format_class(self.shape, dtype=self.dtype)
-        mtx.data = self.group["data"]
-        mtx.indices = self.group["indices"]
+        mtx.data = self._data
+        mtx.indices = self._indices
         mtx.indptr = self._indptr
         return mtx
 
-    def to_memory(self) -> ss.csr_matrix | ss.csc_matrix | SpArray:
+    def to_memory(self) -> CSMatrix | CSArray:
         format_class = get_memory_class(
-            self.format, use_sparray_in_io=settings.shall_use_sparse_array_on_read
+            self.format, use_sparray_in_io=settings.use_sparse_array_on_read
         )
         mtx = format_class(self.shape, dtype=self.dtype)
-        mtx.data = self.group["data"][...]
-        mtx.indices = self.group["indices"][...]
+        mtx.data = self._data[...]
+        mtx.indices = self._indices[...]
         mtx.indptr = self._indptr
         return mtx
 
@@ -613,8 +680,8 @@ def sparse_dataset(group: GroupStorageType) -> abc.CSRDataset | abc.CSCDataset:
 
     >>> import scanpy as sc
     >>> import h5py
-    >>> from anndata import sparse_dataset
-    >>> from anndata import read_elem
+    >>> from anndata.io import sparse_dataset
+    >>> from anndata.io import read_elem
     >>> sc.datasets.pbmc68k_reduced().raw.to_adata().write_h5ad("pbmc.h5ad")
 
     Initialize a sparse dataset from storage
@@ -650,7 +717,8 @@ def sparse_dataset(group: GroupStorageType) -> abc.CSRDataset | abc.CSCDataset:
         return _CSRDataset(group)
     elif encoding_type == "csc":
         return _CSCDataset(group)
-    raise ValueError(f"Unknown encoding type {encoding_type}")
+    msg = f"Unknown encoding type {encoding_type}"
+    raise ValueError(msg)
 
 
 @_subset.register(BaseCompressedSparseDataset)
