@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import inspect
 import warnings
 from collections.abc import Mapping
 from dataclasses import dataclass
@@ -8,8 +9,8 @@ from types import MappingProxyType
 from typing import TYPE_CHECKING, Generic, TypeVar
 
 from anndata._io.utils import report_read_key_on_error, report_write_key_on_error
-from anndata._types import Read, ReadDask, _ReadDaskInternal, _ReadInternal
-from anndata.compat import DaskArray, _read_attr
+from anndata._types import Read, ReadLazy, _ReadInternal, _ReadLazyInternal
+from anndata.compat import DaskArray, ZarrGroup, _read_attr, is_zarr_v2
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Generator, Iterable, Iterator
@@ -23,10 +24,13 @@ if TYPE_CHECKING:
         WriteCallback,
         _WriteInternal,
     )
+    from anndata.experimental.backed._compat import Dataset2D
+    from anndata.experimental.backed._lazy_arrays import CategoricalArray, MaskedArray
     from anndata.typing import RWAble
 
     T = TypeVar("T")
     W = TypeVar("W", bound=_WriteInternal)
+    LazyDataStructures = DaskArray | Dataset2D | CategoricalArray | MaskedArray
 
 
 # TODO: This probably should be replaced by a hashable Mapping due to conversion b/w "_" and "-"
@@ -78,8 +82,8 @@ def write_spec(spec: IOSpec):
     return decorator
 
 
-_R = TypeVar("_R", _ReadInternal, _ReadDaskInternal)
-R = TypeVar("R", Read, ReadDask)
+_R = TypeVar("_R", _ReadInternal, _ReadLazyInternal)
+R = TypeVar("R", Read, ReadLazy)
 
 
 class IORegistry(Generic[_R, R]):
@@ -105,10 +109,11 @@ class IORegistry(Generic[_R, R]):
         if src_type in self.write_specs and (spec != self.write_specs[src_type]):
             # First check for consistency
             current_spec = self.write_specs[src_type]
-            raise TypeError(
+            msg = (
                 "Cannot overwrite IO specifications. Attempted to overwrite encoding "
                 f"for {src_type} from {current_spec} to {spec}"
             )
+            raise TypeError(msg)
         else:
             self.write_specs[src_type] = spec
 
@@ -130,7 +135,6 @@ class IORegistry(Generic[_R, R]):
 
         if dest_type is h5py.File:
             dest_type = h5py.Group
-
         if (dest_type, src_type, modifiers) not in self.write:
             raise IORegistryError._from_write_parts(dest_type, src_type, modifiers)
         internal = self.write[(dest_type, src_type, modifiers)]
@@ -168,7 +172,7 @@ class IORegistry(Generic[_R, R]):
         reader: Reader,
     ) -> R:
         if (src_type, spec, modifiers) not in self.read:
-            raise IORegistryError._from_read_parts("read", self.read, src_type, spec)
+            raise IORegistryError._from_read_parts("read", self.read, src_type, spec)  # noqa: EM101
         internal = self.read[(src_type, spec, modifiers)]
         return partial(internal, _reader=reader)
 
@@ -197,10 +201,8 @@ class IORegistry(Generic[_R, R]):
     ):
         if (src_type, spec, modifiers) in self.read_partial:
             return self.read_partial[(src_type, spec, modifiers)]
-        else:
-            raise IORegistryError._from_read_parts(
-                "read_partial", self.read_partial, src_type, spec
-            )
+        name = "read_partial"
+        raise IORegistryError._from_read_parts(name, self.read_partial, src_type, spec)
 
     def get_spec(self, elem: Any) -> IOSpec:
         if isinstance(elem, DaskArray):
@@ -213,12 +215,13 @@ class IORegistry(Generic[_R, R]):
 
 
 _REGISTRY: IORegistry[_ReadInternal, Read] = IORegistry()
-_LAZY_REGISTRY: IORegistry[_ReadDaskInternal, ReadDask] = IORegistry()
+_LAZY_REGISTRY: IORegistry[_ReadLazyInternal, ReadLazy] = IORegistry()
 
 
 @singledispatch
 def proc_spec(spec) -> IOSpec:
-    raise NotImplementedError(f"proc_spec not defined for type: {type(spec)}.")
+    msg = f"proc_spec not defined for type: {type(spec)}."
+    raise NotImplementedError(msg)
 
 
 @proc_spec.register(IOSpec)
@@ -282,7 +285,7 @@ class Reader:
         return self.callback(read_func, elem.name, elem, iospec=iospec)
 
 
-class DaskReader(Reader):
+class LazyReader(Reader):
     @report_read_key_on_error
     def read_elem(
         self,
@@ -290,16 +293,27 @@ class DaskReader(Reader):
         modifiers: frozenset[str] = frozenset(),
         chunks: tuple[int, ...] | None = None,
         reopen: None | Callable[[], Iterator[StorageType]] = None,
-    ) -> DaskArray:
+        **kwargs,
+    ) -> LazyDataStructures:
         """Read a dask element from a store. See exported function for more details."""
 
         iospec = get_spec(elem)
-        read_func: ReadDask = self.registry.get_read(
+        read_func: ReadLazy = self.registry.get_read(
             type(elem), iospec, modifiers, reader=self
         )
         if self.callback is not None:
             msg = "Dask reading does not use a callback. Ignoring callback."
             warnings.warn(msg, stacklevel=2)
+        read_params = inspect.signature(read_func).parameters
+        for kwarg in kwargs:
+            if kwarg not in read_params:
+                msg = (
+                    f"Keyword argument {kwarg} passed to read_elem_lazy are not supported by the "
+                    "registered read function."
+                )
+                raise ValueError(msg)
+        if "chunks" in read_params:
+            kwargs["chunks"] = chunks
         return read_func(elem, chunks=chunks, reopen=reopen)
 
 
@@ -333,20 +347,31 @@ class Writer:
 
         import h5py
 
+        # we allow stores to have a prefix like /uns which are then written to with keys like /uns/foo
+        if "/" in k.split(store.name)[-1][1:]:
+            msg = "Forward slashes are not allowed in keys."
+            raise ValueError(msg)
+
         if isinstance(store, h5py.File):
             store = store["/"]
 
         dest_type = type(store)
 
-        if elem is None:
-            return lambda *_, **__: None
-
         # Normalize k to absolute path
-        if not PurePosixPath(k).is_absolute():
+        if (
+            (isinstance(store, ZarrGroup) and is_zarr_v2())
+            or isinstance(store, h5py.Group)
+            and not PurePosixPath(k).is_absolute()
+        ):
             k = str(PurePosixPath(store.name) / k)
 
         if k == "/":
-            store.clear()
+            if isinstance(store, ZarrGroup) and not is_zarr_v2():
+                from zarr.core.sync import sync
+
+                sync(store.store.clear())
+            else:
+                store.clear()
         elif k in store:
             del store[k]
 
@@ -379,11 +404,12 @@ def read_elem(elem: StorageType) -> RWAble:
     return Reader(_REGISTRY).read_elem(elem)
 
 
-def read_elem_as_dask(
+def read_elem_lazy(
     elem: StorageType,
     chunks: tuple[int, ...] | None = None,
     reopen: None | Callable[[], Iterator[StorageType]] = None,
-) -> DaskArray:
+    **kwargs,
+) -> LazyDataStructures:
     """
     Read an element from a store lazily.
 
@@ -398,13 +424,68 @@ def read_elem_as_dask(
     chunks, optional
        length `n`, the same `n` as the size of the underlying array.
        Note that the minor axis dimension must match the shape for sparse.
+       Defaults to `(1000, adata.shape[1])` for CSR sparse,
+       `(adata.shape[0], 1000)` for CSC sparse,
+       and the on-disk chunking otherwise for dense.
+       Can use `-1` or `None` to indicate use of the size of the corresponding dimension.
     reopen, optional
         A custom function for re-opening your store in the dask reader.
+
     Returns
     -------
-        DaskArray
+        A "lazy" elem
+
+    Examples
+    --------
+
+    Setting up our example:
+
+    >>> from scanpy.datasets import pbmc3k
+    >>> import tempfile
+    >>> import anndata as ad
+    >>> import zarr
+
+    >>> tmp_path = tempfile.gettempdir()
+    >>> zarr_path = tmp_path + "/adata.zarr"
+
+    >>> adata = pbmc3k()
+    >>> adata.layers["dense"] = adata.X.toarray()
+    >>> adata.write_zarr(zarr_path)
+
+    Reading a sparse matrix from a zarr store lazily, with custom chunk size and default:
+
+    >>> g = zarr.open(zarr_path)
+    >>> adata.X = ad.experimental.read_elem_lazy(g["X"])
+    >>> adata.X
+    dask.array<make_dask_chunk, shape=(2700, 32738), dtype=float32, chunksize=(1000, 32738), chunktype=scipy.csr_matrix>
+    >>> adata.X = ad.experimental.read_elem_lazy(g["X"], chunks=(500, adata.shape[1]))
+    >>> adata.X
+    dask.array<make_dask_chunk, shape=(2700, 32738), dtype=float32, chunksize=(500, 32738), chunktype=scipy.csr_matrix>
+
+    Reading a dense matrix from a zarr store lazily:
+
+    >>> adata.layers["dense"] = ad.experimental.read_elem_lazy(g["layers/dense"])
+    >>> adata.layers["dense"]
+    dask.array<from-zarr, shape=(2700, 32738), dtype=float32, chunksize=(169, 2047), chunktype=numpy.ndarray>
+
+    Making a new anndata object from on-disk, with custom chunks:
+
+    >>> adata = ad.AnnData(
+    ...     obs=ad.io.read_elem(g["obs"]),
+    ...     var=ad.io.read_elem(g["var"]),
+    ...     uns=ad.io.read_elem(g["uns"]),
+    ...     obsm=ad.io.read_elem(g["obsm"]),
+    ...     varm=ad.io.read_elem(g["varm"]),
+    ... )
+    >>> adata.X = ad.experimental.read_elem_lazy(g["X"], chunks=(500, adata.shape[1]))
+    >>> adata.layers["dense"] = ad.experimental.read_elem_lazy(g["layers/dense"])
+
+    We also support using -1 and None as a chunk size to signify the reading the whole axis:
+
+    >>> adata.X = ad.experimental.read_elem_lazy(g["X"], chunks=(500, -1))
+    >>> adata.X = ad.experimental.read_elem_lazy(g["X"], chunks=(500, None))
     """
-    return DaskReader(_LAZY_REGISTRY).read_elem(elem, chunks=chunks, reopen=reopen)
+    return LazyReader(_LAZY_REGISTRY).read_elem(elem, chunks=chunks, reopen=reopen)
 
 
 def write_elem(
