@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import inspect
 import os
-import sys
 import textwrap
 import warnings
 from collections.abc import Iterable
@@ -13,9 +12,6 @@ from functools import partial
 from inspect import Parameter, signature
 from types import GenericAlias
 from typing import TYPE_CHECKING, Generic, NamedTuple, TypeVar, cast
-
-from anndata.compat import CAN_USE_SPARSE_ARRAY
-from anndata.compat.exceptiongroups import add_note
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Sequence
@@ -53,27 +49,14 @@ def describe(self: RegisteredOption, *, as_rst: bool = False) -> str:
     return textwrap.dedent(doc)
 
 
-if sys.version_info >= (3, 11):
+class RegisteredOption(NamedTuple, Generic[T]):
+    option: str
+    default_value: T
+    description: str
+    validate: Callable[[T], None]
+    type: object
 
-    class RegisteredOption(NamedTuple, Generic[T]):
-        option: str
-        default_value: T
-        description: str
-        validate: Callable[[T], None]
-        type: object
-
-        describe = describe
-
-else:
-
-    class RegisteredOption(NamedTuple):
-        option: str
-        default_value: T
-        description: str
-        validate: Callable[[T], None]
-        type: object
-
-        describe = describe
+    describe = describe
 
 
 def check_and_get_environ_var(
@@ -123,6 +106,15 @@ def check_and_get_bool(option, default_value):
         str(int(default_value)),
         ["0", "1"],
         lambda x: bool(int(x)),
+    )
+
+
+def check_and_get_int(option, default_value):
+    return check_and_get_environ_var(
+        f"ANNDATA_{option.upper()}",
+        str(int(default_value)),
+        None,
+        lambda x: int(x),
     )
 
 
@@ -235,7 +227,7 @@ class SettingsManager:
         try:
             validate(default_value)
         except (ValueError, TypeError) as e:
-            add_note(e, f"for option {repr(option)}")
+            e.add_note(f"for option {option!r}")
             raise e
         option_type = type(default_value) if option_type is None else option_type
         self._registered_options[option] = RegisteredOption(
@@ -274,7 +266,7 @@ class SettingsManager:
             ]
         )
         # Update docstring for `SettingsManager.override` as well.
-        doc = cast(str, self.override.__doc__)
+        doc = cast("str", self.override.__doc__)
         insert_index = doc.find("\n        Yields")
         option_docstring = "\t" + "\t".join(
             self.describe(option, should_print_description=False).splitlines(
@@ -329,8 +321,8 @@ class SettingsManager:
         """
         if option in self._deprecated_options:
             deprecated = self._deprecated_options[option]
-            msg = f"{repr(option)} will be removed in {deprecated.removal_version}. {deprecated.message}"
-            warnings.warn(msg, DeprecationWarning)
+            msg = f"{option!r} will be removed in {deprecated.removal_version}. {deprecated.message}"
+            warnings.warn(msg, FutureWarning)
         if option in self._config:
             return self._config[option]
         msg = f"{option} not found."
@@ -395,16 +387,24 @@ settings = SettingsManager()
 ##################################################################################
 # PLACE REGISTERED SETTINGS HERE SO THEY CAN BE PICKED UP FOR DOCSTRING CREATION #
 ##################################################################################
+V = TypeVar("V")
 
 
-def validate_bool(val: Any) -> None:
-    if not isinstance(val, bool):
-        msg = f"{val} not valid boolean"
-        raise TypeError(msg)
+def gen_validator(_type: type[V]) -> Callable[[V], None]:
+    def validate_type(val: V) -> None:
+        if not isinstance(val, _type):
+            msg = f"{val} not valid {_type}"
+            raise TypeError(msg)
+
+    return validate_type
+
+
+validate_bool = gen_validator(bool)
+validate_int = gen_validator(int)
 
 
 settings.register(
-    "shall_remove_unused_categories",
+    "remove_unused_categories",
     default_value=True,
     description="Whether or not to remove unused categories with :class:`~pandas.Categorical`.",
     validate=validate_bool,
@@ -412,7 +412,7 @@ settings.register(
 )
 
 settings.register(
-    "shall_check_uniqueness",
+    "check_uniqueness",
     default_value=True,
     description=(
         "Whether or not to check uniqueness of the `obs` indices on `__init__` of :class:`~anndata.AnnData`."
@@ -430,23 +430,47 @@ settings.register(
 )
 
 
-def validate_sparse_settings(val: Any) -> None:
-    validate_bool(val)
-    if not CAN_USE_SPARSE_ARRAY and cast(bool, val):
-        msg = (
-            "scipy.sparse.cs{r,c}array is not available in current scipy version. "
-            "Falling back to scipy.sparse.cs{r,c}_matrix for reading."
-        )
+def validate_zarr_write_format(format: int):
+    validate_int(format)
+    if format not in {2, 3}:
+        msg = "non-v2 zarr on-disk format not supported"
         raise ValueError(msg)
 
 
 settings.register(
-    "shall_use_sparse_array_on_read",
+    "zarr_write_format",
+    default_value=2,
+    description="Which version of zarr to write to.",
+    validate=validate_zarr_write_format,
+    get_from_env=lambda name, default: check_and_get_environ_var(
+        f"ANNDATA_{name.upper()}",
+        str(default),
+        ["2", "3"],
+        lambda x: int(x),
+    ),
+)
+
+
+def validate_sparse_settings(val: Any) -> None:
+    validate_bool(val)
+
+
+settings.register(
+    "use_sparse_array_on_read",
     default_value=False,
     description="Whether or not to use :class:`scipy.sparse.sparray` as the default class when reading in data",
-    validate=validate_sparse_settings,
+    validate=validate_bool,
     get_from_env=check_and_get_bool,
 )
+
+settings.register(
+    "min_rows_for_chunked_h5_copy",
+    default_value=1000,
+    description="Minimum number of rows at a time to copy when writing out an H5 Dataset to a new location",
+    validate=validate_int,
+    get_from_env=check_and_get_int,
+)
+
 
 ##################################################################################
 ##################################################################################
