@@ -1,21 +1,19 @@
 from __future__ import annotations
 
-import warnings
 from types import MappingProxyType
 from typing import TYPE_CHECKING
 
-import anndata
-
-from .._io.specs import IOSpec, read_elem_as_dask
+from .._io.specs import IOSpec, read_elem_lazy
 from .._types import Read, ReadCallback, StorageType, Write, WriteCallback
+from ..utils import module_get_attr_redirect
 from ._dispatch_io import read_dispatched, write_dispatched
+from .backed import read_lazy
 from .merge import concat_on_disk
 from .multi_files import AnnCollection
 from .pytorch import AnnLoader
 
 if TYPE_CHECKING:
     from typing import Any
-
 
 # Map old name in `anndata.experimental` to new name in `anndata`
 _DEPRECATED = MappingProxyType(
@@ -24,9 +22,9 @@ _DEPRECATED = MappingProxyType(
         for kv in (
             ("CSRDataset", "abc.CSRDataset"),
             ("CSCDataset", "abc.CSCDataset"),
-            "sparse_dataset",
-            "read_elem",
-            "write_elem",
+            ("sparse_dataset", "io.sparse_dataset"),
+            ("read_elem", "io.read_elem"),
+            ("write_elem", "io.write_elem"),
             ("RWAble", "typing.AxisStorable"),
             ("InMemoryElem", "typing.RWAble"),
         )
@@ -35,31 +33,21 @@ _DEPRECATED = MappingProxyType(
 
 
 def __getattr__(attr_name: str) -> Any:
-    if new_path := _DEPRECATED.get(attr_name):
-        msg = (
-            f"Importing {attr_name} from `anndata.experimental` is deprecated. "
-            f"Import anndata.{new_path} instead."
-        )
-        warnings.warn(msg, FutureWarning)
-        # hacky import_object_by_name, but we test all these
-        mod = anndata
-        while "." in new_path:
-            mod_name, new_path = new_path.split(".", 1)
-            mod = getattr(mod, mod_name)
-        return getattr(mod, new_path)
-    msg = f"module {__name__!r} has no attribute {attr_name!r}"
-    raise AttributeError(msg)
+    return module_get_attr_redirect(
+        attr_name, deprecated_mapping=_DEPRECATED, old_module_path="experimental"
+    )
 
 
 __all__ = [
     "AnnCollection",
     "AnnLoader",
-    "read_elem_as_dask",
+    "read_elem_lazy",
     "read_dispatched",
     "write_dispatched",
     "IOSpec",
     "concat_on_disk",
     "Read",
+    "read_lazy",
     "Write",
     "ReadCallback",
     "WriteCallback",
