@@ -10,6 +10,7 @@ import pytest
 from scipy import sparse
 
 import anndata as ad
+from anndata.compat import CSMatrix
 from anndata.tests.helpers import assert_equal, gen_adata
 
 
@@ -39,7 +40,7 @@ def test_sparse_to_dense_disk(tmp_path, mtx_format, to_convert):
     dense_from_mem_pth = tmp_path / "dense_mem.h5ad"
     dense_from_disk_pth = tmp_path / "dense_disk.h5ad"
     mem = gen_adata((50, 50), mtx_format)
-    mem.raw = mem
+    mem.raw = mem.copy()
 
     mem.write_h5ad(mem_pth)
     disk = ad.read_h5ad(mem_pth, backed="r")
@@ -66,7 +67,7 @@ def test_sparse_to_dense_disk(tmp_path, mtx_format, to_convert):
 def test_sparse_to_dense_inplace(tmp_path, spmtx_format):
     pth = tmp_path / "adata.h5ad"
     orig = gen_adata((50, 50), spmtx_format)
-    orig.raw = orig
+    orig.raw = orig.copy()
     orig.write(pth)
     backed = ad.read_h5ad(pth, backed="r+")
     backed.write(as_dense=("X", "raw/X"))
@@ -97,10 +98,10 @@ def test_sparse_to_dense_errors(tmp_path):
 def test_dense_to_sparse_memory(tmp_path, spmtx_format, to_convert):
     dense_path = tmp_path / "dense.h5ad"
     orig = gen_adata((50, 50), np.array)
-    orig.raw = orig
+    orig.raw = orig.copy()
     orig.write_h5ad(dense_path)
-    assert not isinstance(orig.X, sparse.spmatrix)
-    assert not isinstance(orig.raw.X, sparse.spmatrix)
+    assert not isinstance(orig.X, CSMatrix)
+    assert not isinstance(orig.raw.X, CSMatrix)
 
     curr = ad.read_h5ad(dense_path, as_sparse=to_convert, as_sparse_fmt=spmtx_format)
 
