@@ -29,7 +29,8 @@ type_3 = list[int]
 
 def validate_int_list(val) -> bool:
     if not isinstance(val, list) or not [isinstance(type(e), int) for e in val]:
-        raise TypeError(f"{repr(val)} is not a valid int list")
+        msg = f"{val!r} is not a valid int list"
+        raise TypeError(msg)
     return True
 
 
@@ -157,7 +158,7 @@ def test_deprecation(settings: SettingsManager):
     )
     assert described_option.endswith(warning)
     with pytest.warns(
-        DeprecationWarning,
+        FutureWarning,
         match=r"'test_var' will be removed in 0\.1\.0\. This is a deprecation warning!",
     ):
         assert getattr(settings, option) == default_val
@@ -241,18 +242,5 @@ def test_check_and_get_bool_enum(monkeypatch: pytest.MonkeyPatch):
         ),
     ],
 )
-def test_describe(as_rst: bool, expected: str, settings: SettingsManager):
+def test_describe(*, as_rst: bool, expected: str, settings: SettingsManager):
     assert settings.describe("test_var_3", as_rst=as_rst) == expected
-
-
-def test_shall_use_sparse_array_on_read():
-    import anndata as ad
-
-    if not ad.compat.CAN_USE_SPARSE_ARRAY:
-        with pytest.raises(
-            ValueError,
-            match=r"scipy.sparse.cs{r,c}array is not available in current scipy version",
-        ):
-            ad.settings.shall_use_sparse_array_on_read = True
-    else:
-        ad.settings.shall_use_sparse_array_on_read = True
