@@ -1,6 +1,8 @@
 from __future__ import annotations
 
+from contextlib import nullcontext
 from itertools import combinations, product
+from typing import TYPE_CHECKING
 
 import numpy as np
 import pytest
@@ -9,13 +11,8 @@ import anndata as ad
 from anndata import AnnData
 from anndata.tests.helpers import gen_vstr_recarray
 
-
-@pytest.fixture(params=["h5ad", "zarr"])
-def diskfmt(request):
-    return request.param
-
-
-diskfmt2 = diskfmt
+if TYPE_CHECKING:
+    from typing import Literal
 
 
 def assert_str_contents_equal(A, B):
@@ -30,7 +27,9 @@ def assert_str_contents_equal(A, B):
     assert lA == lB
 
 
-def test_io(tmp_path, diskfmt, diskfmt2):
+def test_io(
+    tmp_path, diskfmt: Literal["zarr", "h5ad"], diskfmt2: Literal["zarr", "h5ad"]
+):
     read1 = lambda pth: getattr(ad, f"read_{diskfmt}")(pth)
     write1 = lambda adata, pth: getattr(adata, f"write_{diskfmt}")(pth)
     read2 = lambda pth: getattr(ad, f"read_{diskfmt2}")(pth)
@@ -46,17 +45,24 @@ def test_io(tmp_path, diskfmt, diskfmt2):
     initial = AnnData(np.zeros((3, 3)))
     initial.uns = dict(str_rec=str_recarray, u_rec=u_recarray, s_rec=s_recarray)
 
-    write1(initial, filepth1)
-    disk_once = read1(filepth1)
-    write2(disk_once, filepth2)
-    disk_twice = read2(filepth2)
+    with (
+        pytest.raises(
+            NotImplementedError, match=r"zarr v3 does not support structured dtypes"
+        )
+        if diskfmt == "zarr" and ad.settings.zarr_write_format == 3
+        else nullcontext()
+    ):
+        write1(initial, filepth1)
+        disk_once = read1(filepth1)
+        write2(disk_once, filepth2)
+        disk_twice = read2(filepth2)
 
-    adatas = [initial, disk_once, disk_twice]
-    keys = [
-        "str_rec",
-        "u_rec",
-        # "s_rec"
-    ]
+        adatas = [initial, disk_once, disk_twice]
+        keys = [
+            "str_rec",
+            "u_rec",
+            # "s_rec"
+        ]
 
-    for (ad1, key1), (ad2, key2) in combinations(product(adatas, keys), 2):
-        assert_str_contents_equal(ad1.uns[key1], ad2.uns[key2])
+        for (ad1, key1), (ad2, key2) in combinations(product(adatas, keys), 2):
+            assert_str_contents_equal(ad1.uns[key1], ad2.uns[key2])
